% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{addExperiments}
\alias{addExperiments}
\title{Add a parameter estimation experiment}
\usage{
addExperiments(copasi_exp, model = getCurrentModel())
}
\arguments{
\item{copasi_exp}{object as returned by \code{defineExperiment}}

\item{model}{a model object}
}
\description{
Add a parameter estimation experiment
}
