% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{addOptimizationParameter}
\alias{addOptimizationParameter}
\title{Add an optimization parameter}
\usage{
addOptimizationParameter(copasi_parm, model = getCurrentModel())
}
\arguments{
\item{copasi_parm}{object as returned by \code{defineOptimizationParameter}}

\item{model}{a model object}
}
\description{
Add an optimization parameter
}
