% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{addParameterEstimationParameter}
\alias{addParameterEstimationParameter}
\title{Add a parameter estimation parameter}
\usage{
addParameterEstimationParameter(..., model = getCurrentModel())
}
\arguments{
\item{...}{objects as returned by \code{\link{defineParameterEstimationParameter}}.
Alternatively, the same parameters as used by \code{\link{defineParameterEstimationParameter}}.}

\item{model}{a model object}
}
\description{
Add a parameter estimation parameter
}
\seealso{
\code{\link{defineParameterEstimationParameter}} \code{\link{clearParameterEstimationParameters}}

Other parameter estimation: \code{\link{addExperiments}},
  \code{\link{clearExperiments}},
  \code{\link{clearParameterEstimationParameters}},
  \code{\link{clearValidations}},
  \code{\link{defineExperiments}},
  \code{\link{getParameterEstimationSettings}},
  \code{\link{runParameterEstimation}},
  \code{\link{setParameterEstimationSettings}}
}
\concept{parameter estimation}
