% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{applyInitialState}
\alias{applyInitialState}
\title{Apply the model's initial state}
\usage{
applyInitialState(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{applyInitialState} applies the given model's initial state.
}
\seealso{
Other model actions: 
\code{\link{compileModel}()},
\code{\link{convertToIrreversible}()},
\code{\link{updateInitialState}()}
}
\concept{model actions}
