% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoplot.copasi_ts}
\alias{autoplot.copasi_ts}
\title{Autoplot method for copasi timeseries objects.}
\usage{
autoplot.copasi_ts(object, ...)
}
\arguments{
\item{object}{A copasi timeseries object}

\item{\dots}{Column names selected for plotting}
}
\value{
A ggplot2 plot
}
\description{
Uses ggplot2 to plot timeseries.
The plot has a geom_line layer.
}
