% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biomodels_url}
\alias{biomodels_url}
\title{Generate a URL to a BioModels model.}
\usage{
biomodels_url(id, version = NA, format = c("sbml", "omex"))
}
\arguments{
\item{id}{The numeric BioModels id, as count.}

\item{version}{The version of the model, as count.}

\item{format}{The desired file format for the model.}
}
\value{
A URL to access the BioModels model file.
}
\description{
\code{biomodels_url} generates a URL to a BioModel model file of a specified format.
}
