% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{clearExperiments}
\alias{clearExperiments}
\title{Clear all parameter estimation experiments}
\usage{
clearExperiments(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
Clear all parameter estimation experiments
}
