% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{clearOptimizationParameters}
\alias{clearOptimizationParameters}
\title{Clear all optimization parameters}
\usage{
clearOptimizationParameters(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
Clear all optimization parameters
}
\seealso{
\code{\link{addOptimizationParameter}} \code{\link{defineOptimizationParameter}}

Other optimization: \code{\link{addOptimizationParameter}},
  \code{\link{defineOptimizationParameter}},
  \code{\link{getOptimizationSettings}},
  \code{\link{runOptimization}},
  \code{\link{setOptimizationSettings}}
}
\concept{optimization}
