% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{clearParameters}
\alias{clearParameters}
\title{Clear all parameter estimation parameters}
\usage{
clearParameters(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
Clear all parameter estimation parameters
}
