% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{compartment}
\alias{compartment}
\title{Identify compartments by name}
\usage{
compartment(key = "", reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a string to identify compartments.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of compartment identifiers or references.
}
\description{
\code{compartment} identifies compartments matching a given name fragments.
}
