% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{compartment_strict}
\alias{compartment_strict}
\title{Identify single compartments by name}
\usage{
compartment_strict(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify compartments.}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of compartment identifiers or references.
}
\description{
\code{compartment_strict} identifies strictly one compartment per given name fragment.
}
