% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{compileModel}
\alias{compileModel}
\title{Compile the given model.}
\usage{
compileModel(model = getCurrentModel())
}
\description{
Forces compilation of the given model
}
\seealso{
Other model handlers: \code{\link{getCurrentModel}},
  \code{\link{getLoadedModels}}, \code{\link{openCopasi}},
  \code{\link{setCurrentModel}}
}
\concept{model handlers}
