% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{convertToIrreversible}
\alias{convertToIrreversible}
\title{Convert the model to irreversible reactions}
\usage{
convertToIrreversible(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{convertToIrreversible} converts the given model's reactions to irreversible reactions.
}
\seealso{
Other model actions: \code{\link{applyInitialState}},
  \code{\link{compileModel}},
  \code{\link{updateInitialState}}
}
\concept{model actions}
