% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_copasi_key.R
\docType{data}
\name{copasi_object_types}
\alias{copasi_object_types}
\title{copasi_key <- function(x) {
  assert_that(is_character(x))
  structure(x, class = c("copasi_key", is(x)))
}}
\format{An object of class \code{character} of length 6.}
\usage{
copasi_object_types
}
\description{
#' @export
is.copasi_key <- function(x) {
  inherits(x, "copasi_key")
}
}
\details{
#' @export
as.copasi_key <- function(x) {
  copasi_key(as.character(x))
}
#' @export
format.copasi_key <- function(x, ...) {
  descriptors <- stringr::str_match(x, "=([^=]+)s\\[([^\\[]+)\\]$")
  paste0(
    "# A copasi key object: ", descriptors[2L], ": ", descriptors[3L], "\n"
    # "Key value: ", x
  )
}
#' @export
print.copasi_key <-
  function(x, ...) {
    cat(format(x, ...), "\n")
    invisible(x)
  }

#' @export
show.copasi_key <-
  function(x) {
    print(x)
    invisible(x)
  }
#' @importFrom tibble is_vector_s3
#' @export
is_vector_s3.copasi_key <- function(x) {
  TRUE
}

#' @importFrom tibble type_sum
#' @export
type_sum.copasi_key <- function(x) {
  "ckey"
}

#' @importFrom tibble obj_sum
#' @export
type_sum.copasi_key <- function(x) {
  paste0("ckey [", length(x), "]")
}
}
\keyword{datasets}
