% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{defineExperiments}
\alias{defineExperiments}
\title{Define a parameter estimation experiment}
\usage{
defineExperiments(experiment_type = c("Time Course", "Steady State"),
  data = NULL, types = NULL, mappings = NULL, weight_method = NULL,
  filename = NULL)
}
\arguments{
\item{experiment_type}{string}

\item{data}{list of tidy data frames}

\item{types}{character vector}

\item{mappings}{character vector}

\item{weight_method}{string}

\item{filename}{string}
}
\value{
copasi_exp object for input into related functions
}
\description{
Define a parameter estimation experiment
}
\seealso{
\code{\link{addExperiments}} \code{\link{clearExperiments}}

Other parameter estimation: \code{\link{addExperiments}},
  \code{\link{addParameterEstimationParameter}},
  \code{\link{clearExperiments}},
  \code{\link{clearParameterEstimationParameters}},
  \code{\link{getParameterEstimationSettings}},
  \code{\link{runParameterEstimation}},
  \code{\link{setParameterEstimationSettings}}
}
\concept{parameter estimation}
