% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{defineOptimizationParameter}
\alias{defineOptimizationParameter}
\title{Define an optimization parameter}
\usage{
defineOptimizationParameter(ref, start_value = (lower_bound +
  upper_bound)/2, lower_bound = 1e-06, upper_bound = 1e+06)
}
\arguments{
\item{ref}{value reference}

\item{start_value}{start value}

\item{lower_bound}{lower value bound}

\item{upper_bound}{upper value bound}
}
\value{
copasi_parm object for input into related functions
}
\description{
Define an optimization parameter
}
\seealso{
\code{\link{addOptimizationParameter}} \code{\link{clearOptimizationParameters}}

Other optimization: \code{\link{addOptimizationParameter}},
  \code{\link{clearOptimizationParameters}},
  \code{\link{getOptimizationSettings}},
  \code{\link{runOptimization}},
  \code{\link{setOptimizationSettings}}
}
\concept{optimization}
