% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{defineParameter}
\alias{defineParameter}
\title{Define a parameter estimation parameter}
\usage{
defineParameter(key = NULL, lower.bound = 1e-06, upper.bound = 1e+06,
  start.value = (lower.bound + upper.bound)/2)
}
\arguments{
\item{key}{entity key}

\item{lower.bound}{lower value bound}

\item{upper.bound}{upper value bound}

\item{start.value}{start value}
}
\value{
copasi_parm object for input into related functions
}
\description{
Define a parameter estimation parameter
}
