% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteCompartment}
\alias{deleteCompartment}
\title{Delete a compartment}
\usage{
deleteCompartment(key, model = getCurrentModel())
}
\arguments{
\item{key}{compartment keys}

\item{model}{a model object}
}
\description{
Delete a compartment
}
