% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteEvent}
\alias{deleteEvent}
\title{Delete events}
\usage{
deleteEvent(key, model = getCurrentModel())
}
\arguments{
\item{key}{event keys}

\item{model}{a model object}
}
\description{
\code{deleteEvent} deletes events.
}
\details{
Deletion quietly works recursively (deletes all connected entities) to prevent invalid model states.
}
\seealso{
Other event functions: \code{\link{entity_finders}},
  \code{\link{getEvents}}, \code{\link{newEvent}},
  \code{\link{setEvents}}
}
\concept{event functions}
