% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteReaction}
\alias{deleteReaction}
\title{Delete a reaction}
\usage{
deleteReaction(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction keys}

\item{model}{a model object}
}
\description{
Delete a reaction
}
