% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteSpecies}
\alias{deleteSpecies}
\title{Delete a species}
\usage{
deleteSpecies(key, model = getCurrentModel())
}
\arguments{
\item{key}{species keys}

\item{model}{a model object}
}
\description{
Delete a species
}
