% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{entity_finders}
\alias{entity_finders}
\alias{species}
\alias{species_strict}
\alias{quantity}
\alias{quantity_strict}
\alias{compartment}
\alias{compartment_strict}
\alias{reaction}
\alias{reaction_strict}
\alias{parameter}
\alias{parameter_strict}
\alias{event}
\alias{event_strict}
\alias{kinfunction}
\alias{kinfunction_strict}
\title{Entity finders}
\usage{
species(key = "", reference = NULL, model = getCurrentModel())

species_strict(key, reference = NULL, model = getCurrentModel())

quantity(key = "", reference = NULL, model = getCurrentModel())

quantity_strict(key, reference = NULL, model = getCurrentModel())

compartment(key = "", reference = NULL, model = getCurrentModel())

compartment_strict(key, reference = NULL, model = getCurrentModel())

reaction(key = "", reference = NULL, model = getCurrentModel())

reaction_strict(key, reference = NULL, model = getCurrentModel())

parameter(key = "", reference = NULL, model = getCurrentModel())

parameter_strict(key, reference = NULL, model = getCurrentModel())

event(key = "", model = getCurrentModel())

event_strict(key, model = getCurrentModel())

kinfunction(key = "")

kinfunction_strict(key)
}
\arguments{
\item{key}{entity name fragments

  * main varieties: a string.
  
  * `_strict` varieties: a character vector with each entry uniquely matching one entity.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
A character vector of species identifiers or references.

  References are in the form of "\{name\}" or in rare cases in the form of "<CN=...>" and can
  be used in expressions or functions such as \code{\link{getValue}}.
}
\description{
The entity finder family of functions are a set of helpers to
get unique identifiers for the various entities defined in COPASI.
They give a quick overview of available species, compartments, etc.
}
\details{
There are two classes of entity finders.

The main functions are flexible and return a vector of all entities matching a given name fragment.

The `_strict` varieties expect for all input keys to uniquely match one entity and throw errors otherwise.
This ensures that there are always as many identifiers returned as are given to the functions.
This mechanism is also used by CoRC internally to ensure that functions like \code{\link{getSpecies}}
return one row in the output for each key given as parameter.

Whereas the main functions generally won't help separate an entity `a` from and entity `ab`, the
`_strict` varieties will accept a key "a" because it fully matches the entity `a`, which gets preference
over the partial match with `ab`.

The matching mechanism can be tuned using a \code{\link{regex}} mechanism.

The functions can also be used to find value references by giving the `reference` argument.
Various references such as "Value" or "InitialConcentration" are available for some types of entities.

The \href{https://jpahle.github.io/CoRC/articles/entity_management.html}{online article on managing model entities} provides some further context.
}
\seealso{
Other species functions: \code{\link{deleteSpecies}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{newSpecies}},
  \code{\link{setSpecies}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{newSpecies}},
  \code{\link{setSpecies}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}},
  \code{\link{setGlobalQuantities}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}},
  \code{\link{setGlobalQuantities}}

Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{getCompartmentReferences}},
  \code{\link{getCompartments}},
  \code{\link{newCompartment}},
  \code{\link{setCompartments}}

Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{getCompartmentReferences}},
  \code{\link{getCompartments}},
  \code{\link{newCompartment}},
  \code{\link{setCompartments}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}

Other event functions: \code{\link{deleteEvent}},
  \code{\link{getEvents}}, \code{\link{newEvent}},
  \code{\link{setEvents}}

Other event functions: \code{\link{deleteEvent}},
  \code{\link{getEvents}}, \code{\link{newEvent}},
  \code{\link{setEvents}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{compartment functions}
\concept{event functions}
\concept{global quantity functions}
\concept{reaction functions}
\concept{species functions}
