% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getAreaUnit}
\alias{getAreaUnit}
\title{Get the unit of area}
\usage{
getAreaUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of area
}
\description{
\code{getAreaUnit} gets the unit used for area.
}
\seealso{
\code{\link{setAreaUnit}}

Other model property functions: 
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
