% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getCompartmentReferences}
\alias{getCompartmentReferences}
\title{Get compartment references}
\usage{
getCompartmentReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying compartments}

\item{model}{a model object}
}
\value{
a data frame with compartments and associated references
}
\description{
\code{getCompartmentReferences} returns compartment attribute references as a data frame.
}
