% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getCompartments}
\alias{getCompartments}
\title{Get compartments}
\usage{
getCompartments(key = NULL, prettyExpression = TRUE,
  datamodel = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying compartments}

\item{datamodel}{a model object}
}
\value{
a data frame with compartments and associated information
}
\description{
\code{getCompartments} returns compartments as a data frame.
}
