% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getCopasi-deprecated}
\alias{getCopasi-deprecated}
\alias{getCopasi}
\title{Install COPASI libraries}
\usage{
getCopasi(...)
}
\description{
\code{getCopasi} downloads COPASI libraries or retrieves them from given path.
}
\details{
This function is deprecated as COPASI libraries will now always be included during package installation.

No replacement function is needed.
}
\concept{deprecated}
\keyword{internal}
