% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{getCurrentModel}
\alias{getCurrentModel}
\title{Get the currently active model}
\usage{
getCurrentModel()
}
\value{
a model object
}
\description{
\code{getCurrentModel} returns the currently active model.
}
\details{
By default, the currently active model is the last model that has been used.
If working with a single model, the \code{getCurrentModel} function will supply this model to any function call including the \code{model} argument.
This makes the \code{model} argument used by most of this packages functions safe to ignore in most use cases.
}
\seealso{
Other model handlers: \code{\link{getLoadedModels}},
  \code{\link{openCopasi}}, \code{\link{setCurrentModel}}
}
\concept{model handlers}
