% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getEvents}
\alias{getEvents}
\title{Get events}
\usage{
getEvents(key = NULL, raw_expressions = FALSE, model = getCurrentModel())
}
\arguments{
\item{key}{Optionally, a character vector specifying which events to get.}

\item{raw_expressions}{Whether expressions should be raw (not converted to readable format), as flag.}

\item{model}{A model object.}
}
\value{
Events and associated information, as data frame.
\itemize{
  \item \code{$assignment_target} is a list column containing possibly several targets per event.
  \item \code{$assignment_expression} is a list column containing possibly several expressions per event.
}
}
\description{
\code{getEvents} returns events as a data frame.
}
\seealso{
\code{\link{setEvents}}

Other event functions: \code{\link{deleteEvent}},
  \code{\link{entity_finders}}, \code{\link{newEvent}},
  \code{\link{setEvents}}
}
\concept{event functions}
