% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getGlobalQuantities}
\alias{getGlobalQuantities}
\title{Get global quantities}
\usage{
getGlobalQuantities(key = NULL, prettyExpression = TRUE,
  datamodel = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{datamodel}{a model object}
}
\value{
a data frame with global quantities and associated information
}
\description{
\code{getGlobalQuantities} returns global quantities as a data frame.
}
