% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expressions.R
\name{getInitialValue}
\alias{getInitialValue}
\title{Calculate the initial value of an expression or reference}
\usage{
getInitialValue(expression, model = getCurrentModel())
}
\arguments{
\item{expression}{Expressions to calculate as initial expressions, as character, finite numeric, or logical vector.}

\item{model}{a model object}
}
\value{
a numeric vector of initial values
}
\description{
\code{getValue} calculates the initial value of a given expression or reference.
}
\seealso{
\code{\link{getValue}}

Other expression functions: 
\code{\link{getValue}()}
}
\concept{expression functions}
