% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getLengthUnit}
\alias{getLengthUnit}
\title{Get the unit of length}
\usage{
getLengthUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of length
}
\description{
\code{getLengthUnit} gets the unit used for length.
}
\seealso{
\code{\link{setLengthUnit}}

Other model property functions: \code{\link{getAreaUnit}},
  \code{\link{getInitialTime}},
  \code{\link{getLinkMatrix}}, \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setAreaUnit}}, \code{\link{setInitialTime}},
  \code{\link{setLengthUnit}}, \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
