% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lna.R
\name{getLinearNoiseApproximationSettings}
\alias{getLinearNoiseApproximationSettings}
\alias{getLNA}
\title{Get linear noise approximation settings}
\usage{
getLinearNoiseApproximationSettings(model = getCurrentModel())

getLNA(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of metabolic control analysis task settings including method options.
}
\description{
\code{getLinearNoiseApproximationSettings} gets linear noise approximation settings including method options.
}
