% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getLinkMatrix}
\alias{getLinkMatrix}
\title{Get link matrix}
\usage{
getLinkMatrix(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
matrix
}
\description{
Get link matrix
}
