% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{getLoadedModels}
\alias{getLoadedModels}
\title{Get a list of loaded models}
\usage{
getLoadedModels()
}
\value{
a list of model objects
}
\description{
\code{getLoadedModels} returns a list of all loaded models.
}
