% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{getModelName}
\alias{getModelName}
\title{Get a model's name}
\usage{
getModelName(datamodel = getCurrentModel())
}
\arguments{
\item{datamodel}{a model object}
}
\value{
name
}
\description{
\code{getModelName} gives the name of the model.
}
