% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{getOptimizationSettings}
\alias{getOptimizationSettings}
\alias{getOpt}
\title{Set optimization settings}
\usage{
getOptimizationSettings(model = getCurrentModel())

getOpt(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of parameter estimation task settings including method options.
}
\description{
\code{getOptimizationSettings} gets optimization task settings including method options.
}
