% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{getParameterEstimationSettings}
\alias{getParameterEstimationSettings}
\title{Set parameter estimation settings}
\usage{
getParameterEstimationSettings(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of parameter estimation task settings including method options.
}
\description{
\code{getParameterEstimationSettings} gets parameter estimation task settings including method options.
}
