% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getQuantityUnit}
\alias{getQuantityUnit}
\title{Get the unit of quantity}
\usage{
getQuantityUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of quantity
}
\description{
\code{getQuantityUnit} gets the unit used for quantitiy.
}
\seealso{
\code{\link{setQuantityUnit}}

Other model property functions: \code{\link{getInitialTime}},
  \code{\link{getLinkMatrix}}, \code{\link{getModelName}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setInitialTime}}, \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
