% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getReactionMappings}
\alias{getReactionMappings}
\title{Get reaction parameter mappings}
\usage{
getReactionMappings(key, model = getCurrentModel())
}
\arguments{
\item{key}{Identify which reaction to read by specifying it's key, as string.
Also supports fragments of keys, if uniquely matching one reaction.}

\item{model}{A model object.}
}
\value{
Reaction parameter mappings, as list.
}
\description{
Get reaction parameter mappings
}
\seealso{
\code{\link{setReactionMappings}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
