% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getSpecies}
\alias{getSpecies}
\title{Get species}
\usage{
getSpecies(key = NULL, raw_expressions = FALSE, model = getCurrentModel())
}
\arguments{
\item{key}{Optionally, a character vector specifying which species to get.}

\item{raw_expressions}{Whether expressions should be raw (not converted to readable format), as flag.}

\item{model}{A model object.}
}
\value{
Species and associated information, as data frame.
}
\description{
\code{getSpecies} returns species information as a data frame.
}
\seealso{
\code{\link{getSpeciesReferences}} \code{\link{setSpecies}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{newSpecies}}, \code{\link{setSpecies}}
}
\concept{species functions}
