% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getSpeciesReferences}
\alias{getSpeciesReferences}
\title{Get species references}
\usage{
getSpeciesReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{model}{a model object}
}
\value{
a data frame with species and associated references
}
\description{
\code{getSpeciesReferences} returns species attribute references as a data frame.
}
\seealso{
\code{\link{getSpecies}} \code{\link{setSpecies}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}}, \code{\link{getSpecies}},
  \code{\link{newSpecies}}, \code{\link{setSpecies}}
}
\concept{species functions}
