% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{getSteadyStateSettings}
\alias{getSteadyStateSettings}
\title{Get steady state settings}
\usage{
getSteadyStateSettings(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of time course task settings including method options.
}
\description{
\code{getSteadyStateSettings} gets steady state task settings including method options.
}
