% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getStoichiometryMatrix}
\alias{getStoichiometryMatrix}
\title{Get stoichiometry matrix}
\usage{
getStoichiometryMatrix(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
stoichiometry matrix matrix
}
\description{
Get stoichiometry matrix
}
