% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{getTimeUnit}
\alias{getTimeUnit}
\title{Get the unit of time}
\usage{
getTimeUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of time
}
\description{
\code{getTimeUnit} gets the unit used for time.
}
