% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getValidReactionFunctions}
\alias{getValidReactionFunctions}
\title{Get valid function names for reaction}
\usage{
getValidReactionFunctions(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{model}{a model object}
}
\value{
vector of function names
}
\description{
Get valid function names for reaction
}
\seealso{
\code{\link{setReactionFunction}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
