% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expressions.R
\name{getValue}
\alias{getValue}
\title{Calculate the value of an expression or reference}
\usage{
getValue(expression, model = getCurrentModel())
}
\arguments{
\item{expression}{Expressions to calculate, as character, finite numeric, or logical vector.}

\item{model}{a model object}
}
\value{
a numeric vector of values
}
\description{
\code{getValue} calculates the value of a given expression or reference.
}
\seealso{
\code{\link{getInitialValue}}

Other expression functions: 
\code{\link{getInitialValue}()}
}
\concept{expression functions}
