% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRC.R
\name{getVersion}
\alias{getVersion}
\title{Get COPASI version numbers}
\usage{
getVersion()
}
\value{
Numeric vector consisting of the major version, minor version and build numbers.
}
\description{
\code{getVersion} returns the current version numbers of the COPASI binaries.
}
