% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{kinfunction}
\alias{kinfunction}
\title{Identify functions by name}
\usage{
kinfunction(key = "")
}
\arguments{
\item{key}{a string to identify reactions.}
}
\value{
a character vector of kinetic function identifiers or references.
}
\description{
\code{kinfunction} identifies kinetic functions matching a given name fragment.
}
