% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{kinfunction_strict}
\alias{kinfunction_strict}
\title{Identify single function by name}
\usage{
kinfunction_strict(key)
}
\arguments{
\item{key}{a vector of strings to identify reactions.}
}
\value{
a character vector of kinetic function identifiers or references.
}
\description{
\code{kinfunction_strict} identifies strictly one kinetic function per given name fragment.
}
