% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadModel}
\alias{loadModel}
\title{Load a model}
\usage{
loadModel(path)
}
\arguments{
\item{path}{url or path}
}
\value{
a model object
}
\description{
\code{loadModel} loads a model into COPASI and returns a reference to it.
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadSBMLFromString}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBMLToString}},
  \code{\link{saveSBML}}, \code{\link{unloadAllModels}},
  \code{\link{unloadModel}}
}
\concept{model loading}
