% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadModelFromString}
\alias{loadModelFromString}
\title{Load a model from a string}
\usage{
loadModelFromString(model)
}
\arguments{
\item{model}{string}
}
\description{
\code{loadModelFromString} loads a model from a string.
}
\seealso{
Other model loading: 
\code{\link{loadCombineArchive}()},
\code{\link{loadExamples}()},
\code{\link{loadModel}()},
\code{\link{loadSBMLFromString}()},
\code{\link{loadSBML}()},
\code{\link{newModel}()},
\code{\link{saveCombineArchive}()},
\code{\link{saveModelToString}()},
\code{\link{saveModel}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
