% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadSBMLFromString}
\alias{loadSBMLFromString}
\title{Load an SBML model from a string}
\usage{
loadSBMLFromString(sbml)
}
\arguments{
\item{sbml}{string}
}
\description{
\code{loadSBMLFromString} loads an SBML model from a string.
}
\seealso{
Other model loading: 
\code{\link{loadExamples}()},
\code{\link{loadModelFromString}()},
\code{\link{loadModel}()},
\code{\link{loadSBML}()},
\code{\link{newModel}()},
\code{\link{saveModelToString}()},
\code{\link{saveModel}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
