% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newEvent}
\alias{newEvent}
\title{Create a new event}
\usage{
newEvent(name, trigger_expression, fire_at_initial_time = FALSE,
  trigger_must_remain_true = TRUE, priority_expression = NULL,
  delayed = c("no", "assignment", "calculation"),
  delay_expression = NULL, assignment_target = NULL,
  assignment_expression = NULL, model = getCurrentModel())
}
\arguments{
\item{name}{Name to set, as string.}

\item{trigger_expression}{Trigger expression to set, as string, finite numeric, or logical.}

\item{fire_at_initial_time}{Whether to fire at initial time if true, as logical.}

\item{trigger_must_remain_true}{Whether the trigger must remain true, as logical.}

\item{priority_expression}{Priority expression to set, as string, finite numeric, or logical.}

\item{delayed}{Whether the event assignment and / or calculation is to be delayed ("no", "assignment", "calculation"), as string.}

\item{delay_expression}{Delay expression to set, as string, finite numeric, or logical.}

\item{assignment_target}{Assignment target entities (species, compartments, global quantities), as character vector.}

\item{assignment_expression}{Assignment expressions per event to set, as character, finite numeric, or logical vector.}

\item{model}{a model object}
}
\value{
event key
}
\description{
\code{newEvent} creates a new event.
}
\details{
Default initial value is 1.
Arguments priority from lowest to highest is \code{initial_value}, \code{initial_expression}.

The \href{https://jpahle.github.io/CoRC/articles/model_building.html}{online article on model building} provides some further context.
}
\seealso{
Other event functions: \code{\link{deleteEvent}},
  \code{\link{entity_finders}}, \code{\link{getEvents}},
  \code{\link{setEvents}}
}
\concept{event functions}
