% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newGlobalQuantity}
\alias{newGlobalQuantity}
\title{Create a new global quantity}
\usage{
newGlobalQuantity(name, type = c("fixed", "assignment", "ode"),
  initial.value = 1, expression = NULL, model = getCurrentModel())
}
\arguments{
\item{name}{string}

\item{type}{string}

\item{initial.value}{number}

\item{expression}{string}

\item{model}{a model object}
}
\value{
quantity key
}
\description{
Create a new global quantity
}
\seealso{
Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{setGlobalQuantities}}
}
\concept{global quantity functions}
