% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{newModel}
\alias{newModel}
\title{Load an empty model}
\usage{
newModel()
}
\value{
a model object
}
\description{
\code{newModel} creates a new model returns a reference to it.
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}}, \code{\link{loadSBML}},
  \code{\link{saveModelToString}}, \code{\link{saveModel}},
  \code{\link{unloadAllModels}}, \code{\link{unloadModel}}
}
\concept{model loading}
