% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newReaction}
\alias{newReaction}
\title{Create a new reaction}
\usage{
newReaction(scheme, name = scheme, fun = NULL, mappings = NULL,
  model = getCurrentModel())
}
\arguments{
\item{scheme}{string}

\item{name}{string}

\item{fun}{string}

\item{mappings}{named list}

\item{model}{a model object}
}
\value{
reaction key
}
\description{
Create a new reaction
}
