% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newSpecies}
\alias{newSpecies}
\title{Create a new species}
\usage{
newSpecies(name, compartment = NULL, type = c("reactions", "fixed",
  "assignment", "ode"), initial_concentration = 1, expression = NULL,
  model = getCurrentModel())
}
\arguments{
\item{name}{string}

\item{compartment}{compartment key}

\item{type}{string}

\item{initial_concentration}{number}

\item{expression}{string}

\item{model}{a model object}
}
\value{
species key
}
\description{
\code{newSpecies} creates a new species.
}
\seealso{
Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{setSpecies}}
}
\concept{species functions}
