% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{openCopasi}
\alias{openCopasi}
\title{Open the given model in the Copasi UI}
\usage{
openCopasi(readin = FALSE, copasi_loc = "CopasiUI",
  datamodel = pkg_env$curr_dm)
}
\arguments{
\item{readin}{if TRUE, the function waits for Copasi to quit and then reads in the temporary model file, overwriting the give datamodel}

\item{copasi_loc}{location of CopasiUI}

\item{datamodel}{a model object}
}
\description{
Open the given model in the Copasi UI
}
