% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{parameter}
\alias{parameter}
\title{Identify reaction parameters by name}
\usage{
parameter(key = "", reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a string to identify reaction parameters.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of reaction parameter identifiers or references.
}
\description{
\code{parameter} identifies reaction parameters matching a given name fragment.
}
