% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{parameter_strict}
\alias{parameter_strict}
\title{Identify single reaction parameters by name}
\usage{
parameter_strict(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify reaction parameters.}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of reaction parameter identifiers or references.
}
\description{
\code{reaction_strict} identifies strictly one reaction parameter per given name fragment.
}
