% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{quantity}
\alias{quantity}
\title{Identify quantity by name}
\usage{
quantity(key, reference = NULL, datamodel = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify quantities}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{datamodel}{a model object}
}
\value{
a character vector of copasi keys
}
\description{
\code{quantity} identifies global quantities from the given name fragments.
}
