% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{quantity_strict}
\alias{quantity_strict}
\title{Identify single global quantities by name}
\usage{
quantity_strict(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify global quantities.}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of global quantity identifiers or references.
}
\description{
\code{quantity_strict} identifies strictly one global quantity per given name fragment.
}
