% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{reaction}
\alias{reaction}
\title{Identify reaction by name}
\usage{
reaction(key, reference = NULL, datamodel = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify reactions}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{datamodel}{a model object}
}
\value{
a character vector of copasi keys
}
\description{
\code{reaction} identifies reactions from the given name fragments.
}
