% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{reaction_strict}
\alias{reaction_strict}
\title{Identify single reactions by name}
\usage{
reaction_strict(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify reactions.}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of reaction identifiers or references.
}
\description{
\code{reaction_strict} identifies strictly one reaction per given name fragment.
}
