% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{resetInitialTime}
\alias{resetInitialTime}
\title{Reset the model's initial time}
\usage{
resetInitialTime(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\description{
\code{resetInitialTime} resets the initial time of the model.
}
