% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lna.R
\name{runLinearNoiseApproximation}
\alias{runLinearNoiseApproximation}
\alias{runLNA}
\title{Run linear noise approximation}
\usage{
runLinearNoiseApproximation(perform_steady_state_analysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

runLNA(perform_steady_state_analysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{perform_steady_state_analysis}{flag}

\item{executable}{flag}

\item{method}{list of further method arguments and their values.}

\item{model}{A model object.}
}
\value{
A list of results.
\itemize{
  \item \code{$result_ss} can be one of 'notFound', 'found', 'foundEquilibrium', 'foundNegative'.
  \item \code{$result_lna} can be one of 'allNeg', 'nonNegEigenvaluesExist'.
}
}
\description{
\code{runLinearNoiseApproximation} runs linear noise approximation and returns the results in a list.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other linear noise approximation: \code{\link{getLinearNoiseApproximationSettings}},
  \code{\link{setLinearNoiseApproximationSettings}}
}
\concept{linear noise approximation}
