% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_mca.R
\name{runMetabolicControlAnalysis}
\alias{runMetabolicControlAnalysis}
\alias{runMCA}
\title{Run metabolic control analysis}
\usage{
runMetabolicControlAnalysis(perform_steady_state_analysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

runMCA(perform_steady_state_analysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{perform_steady_state_analysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runMetabolicControlAnalysis} runs metabolic control analysis and returns the results in a list.
}
\seealso{
Other metabolic control analysis: \code{\link{getMetabolicControlAnalysisSettings}},
  \code{\link{setMetabolicControlAnalysisSettings}}
}
\concept{metabolic control analysis}
