% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{runOptimization}
\alias{runOptimization}
\alias{runOpt}
\title{Run optimization}
\usage{
runOptimization(expression = NULL, maximize = NULL, subtask = NULL,
  randomizeStartValues = NULL, calculateStatistics = NULL,
  updateModel = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())

runOpt(expression = NULL, maximize = NULL, subtask = NULL,
  randomizeStartValues = NULL, calculateStatistics = NULL,
  updateModel = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{expression}{string}

\item{maximize}{flag}

\item{subtask}{string

Allowed values: steadyState, timeCourse, scan, fluxMode, optimization, parameterFitting, mca, lyap, tssAnalysis, sens, moieties, crosssection, lna, analytics}

\item{randomizeStartValues}{flag}

\item{calculateStatistics}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{parameters}{copasi_param or list of copasi_param objects}

\item{method}{string or list

Set method with a string or list with entry `method`. The list may contain further method args and their values.
Allowed methods: Statistics, DifferentialEvolution, SRES, EvolutionaryProgram, GeneticAlgorithm, GeneticAlgorithmSR, HookeJeeves, LevenbergMarquardt, NelderMead, ParticleSwarm, Praxis, RandomSearch, ScatterSearch, SimulatedAnnealing, SteepestDescent, TruncatedNewton}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runOptimization} runs optimization and returns the results in a list.
}
