% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{runParamEst}
\alias{runParamEst}
\title{Run parameter estimation}
\usage{
runParamEst(randomizeStartValues = NULL, createParameterSets = NULL,
  calculateStatistics = NULL, updateModel = NULL, parameters = NULL,
  experiments = NULL, method = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{calculateStatistics}{boolean}

\item{updateModel}{boolean}

\item{method}{character or list}

\item{datamodel}{a model object}

\item{calculateJacobian}{boolean}

\item{performStabilityAnalysis}{boolean}
}
\value{
a list of results
}
\description{
\code{runParamEst} runs parameter estimation and returns results in a list.
}
