% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{runParameterEstimation}
\alias{runParameterEstimation}
\alias{runPE}
\title{Run parameter estimation}
\usage{
runParameterEstimation(randomizeStartValues = NULL,
  createParameterSets = NULL, calculateStatistics = NULL,
  updateModel = NULL, executable = NULL, parameters = NULL,
  experiments = NULL, method = NULL, model = getCurrentModel())

runPE(randomizeStartValues = NULL, createParameterSets = NULL,
  calculateStatistics = NULL, updateModel = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{randomizeStartValues}{flag}

\item{createParameterSets}{flag}

\item{calculateStatistics}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{parameters}{copasi_param or list of copasi_param objects}

\item{experiments}{copasi_exp or list of copasi_exp objects}

\item{method}{string or list

Set method with a string or a list with the entry `method`. The list may contain further method args and their values.
Allowed methods: Statistics, DifferentialEvolution, SRES, EvolutionaryProgram, GeneticAlgorithm, GeneticAlgorithmSR, HookeJeeves, LevenbergMarquardt, NelderMead, ParticleSwarm, Praxis, RandomSearch, ScatterSearch, SimulatedAnnealing, SteepestDescent, TruncatedNewton}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runParameterEstimation} runs parameter estimation and returns the results in a list.
}
\seealso{
Other parameter estimation: \code{\link{addExperiments}},
  \code{\link{addParameterEstimationParameter}},
  \code{\link{clearExperiments}},
  \code{\link{clearParameterEstimationParameters}},
  \code{\link{defineExperiments}},
  \code{\link{getParameterEstimationSettings}},
  \code{\link{setParameterEstimationSettings}}
}
\concept{parameter estimation}
