% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_sigmapoint.R
\name{runSigmaPoint}
\alias{runSigmaPoint}
\alias{runSP}
\title{Run sigma point method}
\usage{
runSigmaPoint(alpha = 0.5, beta = 2, kappa = 3, var = NULL, experiments,
  mean_fit_as_basis = TRUE, cl = NULL, model = getCurrentModel())

runSP(alpha = 0.5, beta = 2, kappa = 3, var = NULL, experiments,
  mean_fit_as_basis = TRUE, cl = NULL, model = getCurrentModel())
}
\arguments{
\item{alpha}{number}

\item{beta}{number}

\item{kappa}{count}

\item{var}{Measurement variance, as numeric.
Can be a scalar value for variance on all measurements.
A vector for variance per dependent experimental column or all measurements individually.
A matrix for full covariance matrix.}

\item{experiments}{Object of type copasi_exp as returned by \code{\link{defineExperiments}}.
If it contains a single experiment, the values are understood to be mean values.
If it contains multiple experiments, the experiments have to have identical independent columns.}

\item{mean_fit_as_basis}{Whether to initially fit to the mean experimental data and set those fitted parameter values as starting values.
Given as flag.}

\item{cl}{Either object of type 'cluster' as used in the parallel package, or a count of threads to use, or \code{NULL} to automatically use all cores on current machine.}

\item{model}{A model object.}
}
\value{
A list of results.
\code{$fit_task_results} is a list of all parameter estimation results gathered in this assay.
\code{$sp_means} is the result of term (28).
\code{$sp_cov_matrix} is the result of term (29).
\code{$param_bias} is the result of term (33).
\code{$param_cov_matrix} is the result of term (34).
}
\description{
\code{runSigmaPoint} runs parallel parameter estimations via the sigma point method and returns the results in a list.
}
\concept{sigma point method}
