% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimeCourse}
\alias{runTimeCourse}
\title{Run a time course}
\usage{
runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, method = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{duration}{numeric time course duration}

\item{dt}{numeric}

\item{intervals}{integer}

\item{suppressOutputBefore}{boolean}

\item{outputEvents}{boolean}

\item{saveResultInMemory}{boolean}

\item{startInSteadyState}{boolean}

\item{updateModel}{not yet implemented}

\item{method}{character}

\item{datamodel}{a model object}
}
\value{
a data frame with a time column and species concentration columns
}
\description{
\code{runTimeCourse} runs a time course and returns the species data as a data frame
}
