% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveCombineArchive}
\alias{saveCombineArchive}
\title{Save the model as a combine archive}
\usage{
saveCombineArchive(
  filename,
  include_copasi = TRUE,
  include_sbml = TRUE,
  include_data = TRUE,
  include_sedml = TRUE,
  overwrite = FALSE,
  model = getCurrentModel()
)
}
\arguments{
\item{filename}{a path to save to}

\item{include_copasi}{flag}

\item{include_sbml}{flag}

\item{include_data}{flag}

\item{include_sedml}{flag}

\item{overwrite}{is overwriting existing files allowed?}

\item{model}{a model object}
}
\description{
\code{saveCombineArchive} exports the given model as a combine archive .omex file.
}
\seealso{
Other model loading: 
\code{\link{loadCombineArchive}()},
\code{\link{loadExamples}()},
\code{\link{loadModelFromString}()},
\code{\link{loadModel}()},
\code{\link{loadSBMLFromString}()},
\code{\link{loadSBML}()},
\code{\link{newModel}()},
\code{\link{saveModelToString}()},
\code{\link{saveModel}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
