% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{saveCPS}
\alias{saveCPS}
\title{Save a model as a .cps file}
\usage{
saveCPS(filename = model$getFileName(), overwrite = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{filename}{a path to save to}

\item{overwrite}{is overwriting existing files allowed?}

\item{model}{a model object}
}
\description{
\code{saveCPS} saves a model as a .cps file
}
