% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{saveModel}
\alias{saveModel}
\title{Save a model as a .cps file}
\usage{
saveModel(filename = model$getFileName(), overwrite = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{filename}{a path to save to}

\item{overwrite}{is overwriting existing files allowed?}

\item{model}{a model object}
}
\description{
\code{saveModel} saves a model as a .cps file
}
