% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveModelToString}
\alias{saveModelToString}
\title{Save a model to string}
\usage{
saveModelToString(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{saveModelToString} saves a model to a string.
}
