% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveSBML}
\alias{saveSBML}
\title{Save the model as a SBML file}
\usage{
saveSBML(filename = model$getFileName(), level, version, overwrite = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{filename}{a path to save to}

\item{level}{numeric sbml level}

\item{version}{numeric sbml version}

\item{overwrite}{is overwriting existing files allowed?}

\item{model}{a model object}
}
\description{
\code{saveSBML} exports the given model as a SBML .xml file
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}},
  \code{\link{loadSBMLFromString}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBMLToString}},
  \code{\link{unloadAllModels}}, \code{\link{unloadModel}}
}
\concept{model loading}
