% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveSBMLToString}
\alias{saveSBMLToString}
\title{Save the model to an SBML string}
\usage{
saveSBMLToString(level, version, model = getCurrentModel())
}
\arguments{
\item{level}{numeric sbml level}

\item{version}{numeric sbml version}

\item{model}{a model object}
}
\value{
The model SBML file, as string.
}
\description{
\code{saveSBMLToString} returns the given model as an SBML string.
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}},
  \code{\link{loadSBMLFromString}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBML}},
  \code{\link{unloadAllModels}}, \code{\link{unloadModel}}
}
\concept{model loading}
