% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setAreaUnit}
\alias{setAreaUnit}
\title{Set the unit of area}
\usage{
setAreaUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setAreaUnit} sets the unit used for area.
}
\seealso{
\code{\link{getAreaUnit}}

Other model property functions: \code{\link{getAreaUnit}},
  \code{\link{getInitialTime}},
  \code{\link{getLengthUnit}}, \code{\link{getLinkMatrix}},
  \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setInitialTime}},
  \code{\link{setLengthUnit}}, \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
