% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setCompartments}
\alias{setCompartments}
\title{Set compartments}
\usage{
setCompartments(key = NULL, name = NULL, type = NULL,
  initial_size = NULL, expression = NULL, data = NULL,
  model = getCurrentModel())
}
\arguments{
\item{key}{string uniquely identifying compartment}

\item{name}{string giving name to set}

\item{type}{string giving type ("fixed", "assignment", "ode") to set}

\item{initial_size}{number giving initial size to set}

\item{expression}{string giving expression to set}

\item{data}{a data frame as given by \code{\link{getCompartments}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setCompartments} applies given values to compartments of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which compartment to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.
}
\seealso{
\code{\link{getCompartments}} \code{\link{getCompartmentReferences}}

Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{entity_finders}},
  \code{\link{getCompartmentReferences}},
  \code{\link{getCompartments}},
  \code{\link{newCompartment}}
}
\concept{compartment functions}
