% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{setCurrentModel}
\alias{setCurrentModel}
\title{Set the currently active model}
\usage{
setCurrentModel(model)
}
\arguments{
\item{model}{a model object}
}
\value{
invisibly returns the model object
}
\description{
\code{setCurrentModel} sets the given model as the currently active model.
}
\seealso{
Other model handlers: \code{\link{getCurrentModel}},
  \code{\link{getLoadedModels}}, \code{\link{openCopasi}}
}
\concept{model handlers}
