% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setGlobalQuantities}
\alias{setGlobalQuantities}
\title{Set global quantities}
\usage{
setGlobalQuantities(key = NULL, name = NULL, type = NULL,
  initial_value = NULL, expression = NULL, data = NULL,
  model = getCurrentModel())
}
\arguments{
\item{key}{string uniquely identifying global quantity}

\item{name}{string giving name to set}

\item{type}{string giving type ("fixed", "assignment", "ode") to set}

\item{initial_value}{number giving initial value to set}

\item{expression}{string giving expression to set}

\item{data}{a data frame as given by \code{\link{getGlobalQuantities}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setGlobalQuantities} applies given values to global quantities of the model depending on the \code{key} argument
}
\details{
Use the \code{key} argument to specify which global quantity to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.
}
\seealso{
\code{\link{getGlobalQuantities}} \code{\link{getGlobalQuantityReferences}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}}
}
\concept{global quantity functions}
