% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{setInitialTime}
\alias{setInitialTime}
\title{Set the model's initial time}
\usage{
setInitialTime(time, datamodel = getCurrentModel())
}
\arguments{
\item{time}{numeric}

\item{datamodel}{a model object}
}
\description{
\code{setInitialTime} sets the initial time of the model.
}
