% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setLengthUnit}
\alias{setLengthUnit}
\title{Set the unit of length}
\usage{
setLengthUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setLengthUnit} sets the unit used for length.
}
\seealso{
\code{\link{getLengthUnit}}

Other model property functions: \code{\link{getAreaUnit}},
  \code{\link{getInitialTime}},
  \code{\link{getLengthUnit}}, \code{\link{getLinkMatrix}},
  \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setAreaUnit}}, \code{\link{setInitialTime}},
  \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
