% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lna.R
\name{setLinearNoiseApproximationSettings}
\alias{setLinearNoiseApproximationSettings}
\alias{setLNA}
\title{Set linear noise approximation settings}
\usage{
setLinearNoiseApproximationSettings(performSteadyStateAnalysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

setLNA(performSteadyStateAnalysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{performSteadyStateAnalysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\description{
\code{setLinearNoiseApproximation} sets linear noise approximation task settings including method options.
}
