% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_mca.R
\name{setMetabolicControlAnalysisSettings}
\alias{setMetabolicControlAnalysisSettings}
\alias{setMCA}
\title{Set metabolic control analysis settings}
\usage{
setMetabolicControlAnalysisSettings(
  perform_steady_state_analysis = NULL,
  executable = NULL,
  method = NULL,
  model = getCurrentModel()
)

setMCA(
  perform_steady_state_analysis = NULL,
  executable = NULL,
  method = NULL,
  model = getCurrentModel()
)
}
\arguments{
\item{perform_steady_state_analysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\description{
\code{setMetabolicControlAnalysis} sets metabolic control analysis task settings including method options.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other metabolic control analysis: 
\code{\link{getMetabolicControlAnalysisSettings}()},
\code{\link{runMetabolicControlAnalysis}()}
}
\concept{metabolic control analysis}
