% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setModelName}
\alias{setModelName}
\title{Set a model's name}
\usage{
setModelName(name, model = getCurrentModel())
}
\arguments{
\item{name}{the new model name}

\item{model}{a model object}
}
\description{
\code{setModelName} sets the name of the model.
}
\seealso{
\code{\link{getModelName}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
