% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{setOptimizationSettings}
\alias{setOptimizationSettings}
\alias{setOpt}
\title{Set optimization settings}
\usage{
setOptimizationSettings(randomizeStartValues = NULL,
  createParameterSets = NULL, calculateStatistics = NULL,
  updateModel = NULL, executable = NULL, parameters = NULL,
  experiments = NULL, method = NULL, model = getCurrentModel())

setOpt(randomizeStartValues = NULL, createParameterSets = NULL,
  calculateStatistics = NULL, updateModel = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{randomizeStartValues}{flag}

\item{createParameterSets}{flag}

\item{calculateStatistics}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{parameters}{copasi_param or list of copasi_param objects}

\item{experiments}{copasi_exp or list of copasi_exp objects}

\item{method}{string or list}

\item{model}{a model object}
}
\description{
\code{setOptimizationSettings} sets optimization task settings including parameters, experiments and method options.
}
