% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{setParameterEstimationSettings}
\alias{setParameterEstimationSettings}
\alias{setPE}
\title{Set parameter estimation settings}
\usage{
setParameterEstimationSettings(randomize_start_values = NULL,
  create_parameter_sets = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  experiments = NULL, method = NULL, model = getCurrentModel())

setPE(randomize_start_values = NULL, create_parameter_sets = NULL,
  calculate_statistics = NULL, update_model = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{randomize_start_values}{flag}

\item{create_parameter_sets}{flag}

\item{calculate_statistics}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{parameters}{copasi_parm or list of copasi_parm objects}

\item{experiments}{copasi_exp or list of copasi_exp objects}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'Statistics', 'DifferentialEvolution', 'SRES', 'EvolutionaryProgram', 'GeneticAlgorithm', 'GeneticAlgorithmSR', 'HookeJeeves', 'LevenbergMarquardt', 'NelderMead', 'ParticleSwarm', 'Praxis', 'RandomSearch', 'ScatterSearch', 'SimulatedAnnealing', 'SteepestDescent', 'TruncatedNewton'.}

\item{model}{a model object}
}
\description{
\code{setParameterEstimationSettings} sets parameter estimation task settings including parameters, experiments and method options.
}
\seealso{
Other parameter estimation: \code{\link{addExperiments}},
  \code{\link{addParameterEstimationParameter}},
  \code{\link{clearExperiments}},
  \code{\link{clearParameterEstimationParameters}},
  \code{\link{clearValidations}},
  \code{\link{defineExperiments}},
  \code{\link{getParameterEstimationSettings}},
  \code{\link{runParameterEstimation}}
}
\concept{parameter estimation}
