% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setParameters}
\alias{setParameters}
\title{Set reaction parameters}
\usage{
setParameters(key = NULL, name = NULL, value = NULL, mapping = NULL,
  data = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{string uniquely identifying reaction parameter}

\item{name}{string giving name to set}

\item{value}{number giving value to set}

\item{mapping}{string giving key of global quantity to map to.}

\item{data}{a data frame as given by \code{\link{getParameters}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setParameters} applies given values to reaction parameters of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which reaction parameter to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.
}
\seealso{
\code{\link{getParameters}} \code{\link{getParameterReferences}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
