% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{setQuantityUnit}
\alias{setQuantityUnit}
\title{Set the unit of quantity}
\usage{
setQuantityUnit(unit, datamodel = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{datamodel}{a model object}
}
\description{
\code{setQuantityUnit} sets the unit used for quantity.
}
