% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{setReactionFunction}
\alias{setReactionFunction}
\title{Set a reaction function}
\usage{
setReactionFunction(key, fun, mappings = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{fun}{string}

\item{mappings}{list}

\item{model}{a model object}
}
\description{
Set a reaction function
}
