% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(key = NULL, name = NULL, compartment = NULL, type = NULL,
  initial_concentration = NULL, initial_number = NULL, expression = NULL,
  data = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{string uniquely identifying species}

\item{name}{string giving name to set}

\item{compartment}{string giving key of new compartment to set}

\item{type}{string giving type ("fixed", "assignment", "reactions", "ode") to set}

\item{initial_concentration}{number giving initial concentration to set}

\item{initial_number}{number giving initial particle number to set}

\item{expression}{string giving expression to set}

\item{data}{a data frame as given by \code{\link{getSpecies}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setSpecies} applies given values to species of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which species to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.
}
\seealso{
\code{\link{getSpecies}} \code{\link{getSpeciesReferences}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{newSpecies}}
}
\concept{species functions}
