% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{setVolumeUnit}
\alias{setVolumeUnit}
\title{Set the unit of volume}
\usage{
setVolumeUnit(unit, datamodel = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{datamodel}{a model object}
}
\description{
\code{setVolumeUnit} sets the unit used for volume.
}
