% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{species}
\alias{species}
\title{Identify species by name}
\usage{
species(key = "", reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a string to identify species.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of species identifiers or references.
}
\description{
\code{species} identifies species matching a given name fragment.
}
