% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{species_strict}
\alias{species_strict}
\title{Identify single species by name}
\usage{
species_strict(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify species.}

\item{reference}{a string or vector of strings naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of species identifiers or references.
}
\description{
\code{species_strict} identifies strictly one species per given name fragment.
}
