% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{unloadAllModels}
\alias{unloadAllModels}
\title{Unload all loaded models}
\usage{
unloadAllModels()
}
\description{
\code{unloadAllModels} frees memory by unloading all loaded models from copasi
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBML}},
  \code{\link{unloadModel}}
}
\concept{model loading}
