% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{unloadModel}
\alias{unloadModel}
\title{Unload a model}
\usage{
unloadModel(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{unloadModel} frees memory by unloading a model from copasi
}
