% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{updateInitialState}
\alias{updateInitialState}
\title{Update the model's initial state from the current state}
\usage{
updateInitialState(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{updateInitialState} updates the given model's initial state from the current state.
}
\seealso{
Other model actions: 
\code{\link{applyInitialState}()},
\code{\link{compileModel}()},
\code{\link{convertToIrreversible}()}
}
\concept{model actions}
