% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{addOptimizationParameter}
\alias{addOptimizationParameter}
\title{Add an optimization parameter}
\usage{
addOptimizationParameter(copasi_parm, model = getCurrentModel())
}
\arguments{
\item{copasi_parm}{object as returned by \code{\link{defineOptimizationParameter}}}

\item{model}{a model object}
}
\description{
Add an optimization parameter
}
\seealso{
\code{\link{defineOptimizationParameter}} \code{\link{clearOptimizationParameters}}

Other optimization: \code{\link{clearOptimizationParameters}},
  \code{\link{defineOptimizationParameter}},
  \code{\link{getOptimizationSettings}},
  \code{\link{runOptimization}},
  \code{\link{setOptimizationSettings}}
}
\concept{optimization}
