% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{addParameterEstimationParameter}
\alias{addParameterEstimationParameter}
\title{Add a parameter estimation parameter}
\usage{
addParameterEstimationParameter(copasi_parm, model = getCurrentModel())
}
\arguments{
\item{copasi_parm}{object as returned by \code{defineParameterEstimationParameter}}

\item{model}{a model object}
}
\description{
Add a parameter estimation parameter
}
