% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tasks.R
\name{autoplot.copasi_ts}
\alias{autoplot.copasi_ts}
\title{Autoplot method for COPASI timeseries objects.}
\usage{
autoplot.copasi_ts(object, ..., use_concentrations = TRUE)
}
\arguments{
\item{object}{The data to be plotted, as returned by \code{\link{runTimeCourse}}. This object is of type \code{copasi_ts}.}

\item{\dots}{Column names selected for plotting, as strings. Uses partial matching. Also accepts full keys.}

\item{use_concentrations}{Whether to print concentrations or particle numbers for species, as flag.}
}
\value{
A ggplot2 plot
}
\description{
Uses ggplot2 to plot timeseries.
The plot has a geom_line layer.
}
