% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{clearExperiments}
\alias{clearExperiments}
\title{Clear all parameter estimation experiments}
\usage{
clearExperiments(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
Clear all parameter estimation experiments
}
\seealso{
\code{\link{addExperiments}}

Other parameter estimation: 
\code{\link{addExperiments}()},
\code{\link{addParameterEstimationParameter}()},
\code{\link{clearParameterEstimationParameters}()},
\code{\link{clearValidations}()},
\code{\link{defineExperiments}()},
\code{\link{getParameterEstimationSettings}()},
\code{\link{runParameterEstimation}()},
\code{\link{setParameterEstimationSettings}()}
}
\concept{parameter estimation}
