% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{compartment}
\alias{compartment}
\title{Identify compartment by name}
\usage{
compartment(key, reference = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a vector of strings to identify compartments}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{datamodel}{a model object}
}
\value{
a character vector of copasi keys
}
\description{
\code{compartment} identifies compartments from the given name fragments.
}
