% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{compileModel}
\alias{compileModel}
\title{Compile the model}
\usage{
compileModel(force = FALSE, model = getCurrentModel())
}
\arguments{
\item{force}{flag}

\item{model}{a model object}
}
\description{
\code{compileModel} potentially forces compilation of the given model.
}
\seealso{
Other model actions: 
\code{\link{applyInitialState}()},
\code{\link{convertToIrreversible}()},
\code{\link{updateInitialState}()}
}
\concept{model actions}
