% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRC.R
\docType{package}
\name{CoRC-package}
\alias{CoRC}
\alias{CoRC-package}
\title{CoRC: Copasi R Console.}
\description{
Run timecourses and stuff
}
\section{quantity accessors}{

Use dataframes to manipulate species and quantities and stuff
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpahle/CoRC}
  \item Report bugs at \url{https://github.com/jpahle/CoRC/issues}
}

}
\author{
\strong{Maintainer}: Jonas Förster \email{jonasfoe@gmail.com}

}
