% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRC.R
\docType{package}
\name{CoRC-package}
\alias{CoRC}
\alias{CoRC-package}
\title{CoRC: Copasi R Console.}
\description{
CoRC, the Copasi R Connector, links the Complex Pathway Simulator COPASI to R.
It provides easy access to the powerful biochemical model editing, simulation and analysis backend of Copasi.
This allows the user to develop elaborate scripts and workflows for analyses that would require a great deal of tedious manual work otherwise.
These scripts can then be run interactively or be sent to cluster or cloud facilities for more demanding calculations.
}
\details{
CoRC features:
  
* high-level API for Copasi
* define reproducible workflows using Copasi's powerful modelling tools
* easy parallization of Copasi tasks
* enables straight forward publication-ready plotting with the R ecosystem
* ...

It is based on a current development version of Copasi but is expected to closely follow official Copasi releases in the future.
}
\section{quantity accessors}{

Use dataframes to manipulate species and quantities and stuff
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpahle/CoRC}
  \item Report bugs at \url{https://github.com/jpahle/CoRC/issues}
}

}
\author{
\strong{Maintainer}: Jonas Förster \email{jonasfoerster@gmail.com}

Authors:
\itemize{
  \item Jürgen Pahle \email{juergen.pahle@bioquant.uni-heidelberg.de} [funder]
}

}
