% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{createSpecies}
\alias{createSpecies}
\title{Create a new species}
\usage{
createSpecies(name = NULL, compartment = NULL, type = c("fixed",
  "assignment", "reactions", "ode"), initial.concentration = 1,
  expression = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\description{
\code{createSpecies} creates a new species.
}
