% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{defineOptimizationParameter}
\alias{defineOptimizationParameter}
\title{Define an optimization parameter}
\usage{
defineOptimizationParameter(key = NULL, lower.bound = 1e-06,
  upper.bound = 1e+06, start.value = (lower.bound + upper.bound)/2)
}
\arguments{
\item{key}{entity key}

\item{lower.bound}{lower value bound}

\item{upper.bound}{upper value bound}

\item{start.value}{start value}
}
\value{
copasi_parm object for input into related functions
}
\description{
Define an optimization parameter
}
\seealso{
\code{\link{addOptimizationParameter}} \code{\link{clearOptimizationParameters}}

Other optimization: \code{\link{addOptimizationParameter}},
  \code{\link{getOptimizationSettings}},
  \code{\link{runOptimization}},
  \code{\link{setOptimizationSettings}}
}
\concept{optimization}
