% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteCompartment}
\alias{deleteCompartment}
\title{Delete compartments}
\usage{
deleteCompartment(key, model = getCurrentModel())
}
\arguments{
\item{key}{compartment keys}

\item{model}{a model object}
}
\description{
\code{deleteCompartment} deletes compartments.
}
\details{
Deletion quietly works recursively (deletes all connected entities) to prevent invalid model states.
}
\seealso{
Other compartment functions: 
\code{\link{entity_finders}},
\code{\link{getCompartmentReferences}()},
\code{\link{getCompartments}()},
\code{\link{newCompartment}()},
\code{\link{setCompartments}()}
}
\concept{compartment functions}
