% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteGlobalQuantity}
\alias{deleteGlobalQuantity}
\title{Delete global quantities}
\usage{
deleteGlobalQuantity(key, model = getCurrentModel())
}
\arguments{
\item{key}{quantity keys}

\item{model}{a model object}
}
\description{
\code{deleteGlobalQuantity} deletes global quantities.
}
\details{
Deletion quietly works recursively (deletes all connected entities) to prevent invalid model states.
}
\seealso{
Other global quantity functions: 
\code{\link{entity_finders}},
\code{\link{getGlobalQuantities}()},
\code{\link{getGlobalQuantityReferences}()},
\code{\link{newGlobalQuantity}()},
\code{\link{setGlobalQuantities}()}
}
\concept{global quantity functions}
