% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{deleteReaction}
\alias{deleteReaction}
\title{Delete reactions}
\usage{
deleteReaction(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction keys}

\item{model}{a model object}
}
\description{
\code{deleteReaction} deletes reactions.
}
\details{
Deletion quietly works recursively (deletes all connected entities) to prevent invalid model states.
}
\seealso{
Other reaction functions: 
\code{\link{clearCustomKineticFunctions}()},
\code{\link{deleteKineticFunction}()},
\code{\link{entity_finders}},
\code{\link{getParameterReferences}()},
\code{\link{getParameters}()},
\code{\link{getReactionMappings}()},
\code{\link{getReactionReferences}()},
\code{\link{getReactions}()},
\code{\link{getValidReactionFunctions}()},
\code{\link{newKineticFunction}()},
\code{\link{newReaction}()},
\code{\link{setParameters}()},
\code{\link{setReactionFunction}()},
\code{\link{setReactionMappings}()},
\code{\link{setReactions}()}
}
\concept{reaction functions}
