% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getCompartmentReferences}
\alias{getCompartmentReferences}
\title{Get compartment references}
\usage{
getCompartmentReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{Optionally, a character vector specifying which compartments to get.}

\item{model}{A model object.}
}
\value{
Compartments and associated references, as data frame.
}
\description{
\code{getCompartmentReferences} returns compartment attribute references as a data frame.
}
\seealso{
\code{\link{getCompartments}} \code{\link{setCompartments}}

Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{entity_finders}},
  \code{\link{getCompartments}},
  \code{\link{newCompartment}},
  \code{\link{setCompartments}}
}
\concept{compartment functions}
