% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getCompartments}
\alias{getCompartments}
\title{Get compartments}
\usage{
getCompartments(key = NULL, rawExpressions = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying compartments}

\item{model}{a model object}

\item{rawExpression}{a flag on whether expressions should be raw (not converted to readable format)}
}
\value{
a data frame with compartments and associated information
}
\description{
\code{getCompartments} returns compartments as a data frame.
}
