% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getCompartments}
\alias{getCompartments}
\title{Get compartments}
\usage{
getCompartments(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
a data frame with compartments and associated information
}
\description{
\code{getCompartments} returns all compartments as a data frame.
}
