% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRC.R
\name{getCopasi}
\alias{getCopasi}
\title{Install COPASI binaries}
\usage{
getCopasi(path = NULL, force = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{Optionally, a file path to COPASI binaries, as string.}

\item{force}{Whether to force overwriting the binaries, as flag.}

\item{quiet}{Whether to silence messages, as flag.}
}
\description{
\code{getCopasi} automatically downloads binaries or retrieves them from given path.
}
\details{
In case of no internet connection, run the function, retrieve the URL from the error message, and download the file manually.
Alternatively, download the binaries from the release section on github.
Copy the file to a local path and give its path via the \code{path} argument.

To install COPASI binaries, you need write access to the packages installation folder.
}
