% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoRC.R
\name{getCopasi}
\alias{getCopasi}
\title{Install copasi binaries}
\usage{
getCopasi(path = NULL, force = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{optional file path to copasi binaries}

\item{force}{optional bool to force overwriting the binaries}

\item{quiet}{optional bool to silence messages}
}
\description{
\code{getCopasi} automatically downloads binaries or retrieves them from given path.
To install copasi binaries, you need write access to the packages installation folder.
}
