% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{getCurrentModel}
\alias{getCurrentModel}
\title{Get the currently active model}
\usage{
getCurrentModel()
}
\value{
a model object
}
\description{
\code{getCurrentModel} returns the currently active model.
}
\seealso{
Other model handlers: \code{\link{getLoadedModels}},
  \code{\link{openCopasi}}, \code{\link{setCurrentModel}}
}
\concept{model handlers}
