% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{getCurrentModel}
\alias{getCurrentModel}
\title{Get the currently active model}
\usage{
getCurrentModel()
}
\value{
a model object
}
\description{
\code{getCurrentModel} returns the currently active model.
}
