% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getGlobalQuantities}
\alias{getGlobalQuantities}
\title{Get Global Quantities}
\usage{
getGlobalQuantities(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
a data frame with global quantities and associated information
}
\description{
\code{getGlobalQuantities} returns all species as a data frame.
}
