% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getGlobalQuantities}
\alias{getGlobalQuantities}
\title{Get global quantities}
\usage{
getGlobalQuantities(key = NULL, rawExpressions = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{rawExpressions}{a flag on whether expressions should be raw (not converted to readable format)}

\item{model}{a model object}
}
\value{
a data frame with global quantities and associated information
}
\description{
\code{getGlobalQuantities} returns global quantities as a data frame.
}
\seealso{
\code{\link{getGlobalQuantityReferences}} \code{\link{setGlobalQuantities}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}},
  \code{\link{setGlobalQuantities}}
}
\concept{global quantity functions}
