% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getGlobalQuantities}
\alias{getGlobalQuantities}
\title{Get global quantities}
\usage{
getGlobalQuantities(key = NULL, raw_expressions = FALSE,
  model = getCurrentModel())
}
\arguments{
\item{key}{Optionally, a character vector specifying which global quantities to get.}

\item{raw_expressions}{Whether expressions should be raw (not converted to readable format), as flag.}

\item{model}{A model object.}
}
\value{
Global quantities and associated information, as data frame.
}
\description{
\code{getGlobalQuantities} returns global quantities as a data frame.
}
\seealso{
\code{\link{getGlobalQuantityReferences}} \code{\link{setGlobalQuantities}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}},
  \code{\link{setGlobalQuantities}}
}
\concept{global quantity functions}
