% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getGlobalQuantityReferences}
\alias{getGlobalQuantityReferences}
\title{Get global quantitiy references}
\usage{
getGlobalQuantityReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{model}{a model object}
}
\value{
a data frame with global quantities and associated references
}
\description{
\code{getGlobalQuantityReferences} returns global quantity attribute references as a data frame.
}
