% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getInitialTime}
\alias{getInitialTime}
\title{Get the model's initial time}
\usage{
getInitialTime(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{getInitialTime} gets the initial time of the model.
}
\seealso{
\code{\link{setInitialTime}}

Other model property functions: \code{\link{getAreaUnit}},
  \code{\link{getLengthUnit}}, \code{\link{getLinkMatrix}},
  \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setAreaUnit}}, \code{\link{setInitialTime}},
  \code{\link{setLengthUnit}}, \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
