% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{getInitialTime}
\alias{getInitialTime}
\title{Get the model's inital time}
\usage{
getInitialTime(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\description{
\code{getInitialTime} gets the initial time of the model.
}
