% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getLinkMatrix}
\alias{getLinkMatrix}
\title{Get link matrix}
\usage{
getLinkMatrix(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
matrix
}
\description{
Get link matrix
}
\seealso{
Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
