% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_mca.R
\name{getMetabolicControlAnalysisSettings}
\alias{getMetabolicControlAnalysisSettings}
\alias{getMCA}
\title{Get metabolic control analysis settings}
\usage{
getMetabolicControlAnalysisSettings(model = getCurrentModel())

getMCA(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of metabolic control analysis task settings including method options.
}
\description{
\code{getMetabolicControlAnalysisSettings} gets metabolic control analysis settings including method options.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other metabolic control analysis: \code{\link{runMetabolicControlAnalysis}},
  \code{\link{setMetabolicControlAnalysisSettings}}
}
\concept{metabolic control analysis}
