% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getModelName}
\alias{getModelName}
\title{Get the model's name}
\usage{
getModelName(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
name
}
\description{
\code{getModelName} gives the name of the model.
}
\seealso{
\code{\link{setModelName}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
