% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{getParameterEstimationSettings}
\alias{getParameterEstimationSettings}
\alias{getPE}
\title{Set parameter estimation settings}
\usage{
getParameterEstimationSettings(model = getCurrentModel())

getPE(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of parameter estimation task settings including method options.
}
\description{
\code{getParameterEstimationSettings} gets parameter estimation task settings including method options.
}
\seealso{
Other parameter estimation: \code{\link{addExperiments}},
  \code{\link{addParameterEstimationParameter}},
  \code{\link{clearExperiments}},
  \code{\link{clearParameterEstimationParameters}},
  \code{\link{defineExperiments}},
  \code{\link{runParameterEstimation}},
  \code{\link{setParameterEstimationSettings}}
}
\concept{parameter estimation}
