% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getParameterReferences}
\alias{getParameterReferences}
\title{Get reaction parameter references}
\usage{
getParameterReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying reactions parameters}

\item{model}{a model object}
}
\value{
a data frame with reaction parameters and associated references
}
\description{
\code{getParameterReferences} returns reaction parameters as a data frame.
}
