% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getParameters}
\alias{getParameters}
\title{Get reaction parameters}
\usage{
getParameters(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying reactions parameters}

\item{model}{a model object}
}
\value{
a data frame with reaction parameters and associated information
}
\description{
\code{getParameters} returns reaction parameters as a data frame.
}
