% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{getQuantityUnit}
\alias{getQuantityUnit}
\title{Get the unit of quantity}
\usage{
getQuantityUnit(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
unit of quantity
}
\description{
\code{getQuantityUnit} gets the unit used for quantitiy.
}
