% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getReactionMappings}
\alias{getReactionMappings}
\title{Get reaction parameter mappings}
\usage{
getReactionMappings(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{model}{a model object}
}
\value{
list of parameter mappings
}
\description{
Get reaction parameter mappings
}
\seealso{
\code{\link{setReactionMappings}}

Other reaction functions: \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}}, \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
