% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{getReactionMappings}
\alias{getReactionMappings}
\title{Get reaction parameter mappings}
\usage{
getReactionMappings(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{model}{a model object}
}
\value{
list of parameter mappings
}
\description{
Get reaction parameter mappings
}
