% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getReactions}
\alias{getReactions}
\title{Get reactions}
\usage{
getReactions(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
a data frame with reactions and associated information
}
\description{
\code{getReactions} returns all reactions as a data frame.
}
