% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getReactions}
\alias{getReactions}
\title{Get reactions}
\usage{
getReactions(key = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{datamodel}{a model object}
}
\value{
a data frame with reactions and associated information
}
\description{
\code{getReactions} returns reactions as a data frame.
}
