% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getReducedStoichiometryMatrix}
\alias{getReducedStoichiometryMatrix}
\title{Get reduced stoichiometry matrix}
\usage{
getReducedStoichiometryMatrix(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
reduced stoichiometry matrix
}
\description{
Get reduced stoichiometry matrix
}
\seealso{
\code{\link{getStoichiometryMatrix}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
