% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getReducedStoichiometryMatrix}
\alias{getReducedStoichiometryMatrix}
\title{Get reduced stoichiometry matrix}
\usage{
getReducedStoichiometryMatrix(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
reduced stoichiometry matrix
}
\description{
Get reduced stoichiometry matrix
}
