% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getSpecies}
\alias{getSpecies}
\title{Get species}
\usage{
getSpecies(key = NULL, rawExpressions = FALSE, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{rawExpressions}{a flag on whether expressions should be raw (not converted to readable format)}

\item{model}{a model object}
}
\value{
a data frame with species and associated information
}
\description{
\code{getSpecies} returns species information as a data frame.
}
\seealso{
\code{\link{getSpeciesReferences}} \code{\link{setSpecies}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{newSpecies}}, \code{\link{setSpecies}}
}
\concept{species functions}
