% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getSpecies}
\alias{getSpecies}
\title{Get species}
\usage{
getSpecies(datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}
}
\value{
a data frame with species and associated information
}
\description{
\code{getSpecies} returns all species as a data frame.
}
