% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getSpecies}
\alias{getSpecies}
\title{Get species}
\usage{
getSpecies(key = NULL, prettyExpression = TRUE,
  datamodel = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{datamodel}{a model object}
}
\value{
a data frame with species and associated information
}
\description{
\code{getSpecies} returns species information as a data frame.
}
