% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getSpeciesReferences}
\alias{getSpeciesReferences}
\title{Get species references}
\usage{
getSpeciesReferences(key = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{Optionally, a character vector specifying which species to get.}

\item{model}{A model object.}
}
\value{
Species and associated references, as data frame.
}
\description{
\code{getSpeciesReferences} returns species attribute references as a data frame.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/entity_management.html}{online article on managing model entities} provides some further context.
}
\seealso{
\code{\link{getSpecies}} \code{\link{setSpecies}}

Other species functions: 
\code{\link{deleteSpecies}()},
\code{\link{entity_finders}},
\code{\link{getSpecies}()},
\code{\link{newSpecies}()},
\code{\link{setSpecies}()}
}
\concept{species functions}
