% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{getSpeciesReferences}
\alias{getSpeciesReferences}
\title{Get species references}
\usage{
getSpeciesReferences(key = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{datamodel}{a model object}
}
\value{
a data frame with species and associated references
}
\description{
\code{getSpeciesReferences} returns species attribute references as a data frame.
}
