% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{getSteadyStateSettings}
\alias{getSteadyStateSettings}
\alias{getSS}
\title{Get steady state settings}
\usage{
getSteadyStateSettings(model = getCurrentModel())

getSS(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
A list of steady state task settings including method options.
}
\description{
\code{getSteadyStateSettings} gets steady state task settings including method options.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other steady state: 
\code{\link{runSteadyState}()},
\code{\link{setSteadyStateSettings}()}
}
\concept{steady state}
