% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getTimeUnit}
\alias{getTimeUnit}
\title{Get the unit of time}
\usage{
getTimeUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of time
}
\description{
\code{getTimeUnit} gets the unit used for time.
}
\seealso{
\code{\link{setTimeUnit}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
