% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{getValidReactionFunctions}
\alias{getValidReactionFunctions}
\title{Get valid function names for reaction}
\usage{
getValidReactionFunctions(key, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{model}{a model object}
}
\value{
vector of function names
}
\description{
Get valid function names for reaction
}
