% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expressions.R
\name{getValue}
\alias{getValue}
\title{Calculate the value of an expression or reference}
\usage{
getValue(expression, model = getCurrentModel())
}
\arguments{
\item{expression}{a character vector of expressions}

\item{model}{a model object}
}
\value{
a numeric vector of values
}
\description{
\code{getValue} calculates the value of a given expression or reference.
}
