% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getVolumeUnit}
\alias{getVolumeUnit}
\title{Get the unit of volume}
\usage{
getVolumeUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of volume
}
\description{
\code{getTimeUnit} gets the unit used for volume.
}
