% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{getVolumeUnit}
\alias{getVolumeUnit}
\title{Get the unit of volume}
\usage{
getVolumeUnit(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
unit of volume
}
\description{
\code{getVolumeUnit} gets the unit used for volume.
}
\seealso{
\code{\link{setVolumeUnit}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
