% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadExamples}
\alias{loadExamples}
\title{Load example models}
\usage{
loadExamples(indices = NULL)
}
\arguments{
\item{indices}{optional indices of example models to load.}
}
\value{
a list of model objects
}
\description{
\code{loadExamples} loads several example models and returns them in a list.
}
\seealso{
Other model loading: \code{\link{loadModelFromString}},
  \code{\link{loadModel}},
  \code{\link{loadSBMLFromString}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBMLToString}},
  \code{\link{saveSBML}}, \code{\link{unloadAllModels}},
  \code{\link{unloadModel}}
}
\concept{model loading}
