% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{loadExamples}
\alias{loadExamples}
\title{Load example models}
\usage{
loadExamples(indices = NULL)
}
\arguments{
\item{indices}{optional indices of example models to load.}
}
\value{
a list of model objects
}
\description{
\code{loadExamples} loads several example models and returns them in a list.
}
