% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadModel}
\alias{loadModel}
\title{Load a model}
\usage{
loadModel(path)
}
\arguments{
\item{path}{url or path}
}
\value{
a model object
}
\description{
\code{loadModel} loads a model into copasi and returns a reference to it.
}
