% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadModelFromString}
\alias{loadModelFromString}
\title{Load a model from string}
\usage{
loadModelFromString(model)
}
\arguments{
\item{model}{string}
}
\description{
\code{loadModelFromString} loads a model from a string.
}
