% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadSBML}
\alias{loadSBML}
\title{Load SBML data}
\usage{
loadSBML(path)
}
\arguments{
\item{path}{url or path}
}
\value{
a model object
}
\description{
\code{loadSBML} loads SBML data into copasi and returns a reference to it.
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}}, \code{\link{newModel}},
  \code{\link{saveModelToString}}, \code{\link{saveModel}},
  \code{\link{unloadAllModels}}, \code{\link{unloadModel}}
}
\concept{model loading}
