% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{loadSBML}
\alias{loadSBML}
\title{Load an SBML model}
\usage{
loadSBML(path)
}
\arguments{
\item{path}{url or path}
}
\value{
a model object
}
\description{
\code{loadSBML} loads an SBML model into COPASI and returns a reference to it.
}
\seealso{
Other model loading: 
\code{\link{loadCombineArchive}()},
\code{\link{loadExamples}()},
\code{\link{loadModelFromString}()},
\code{\link{loadModel}()},
\code{\link{loadSBMLFromString}()},
\code{\link{newModel}()},
\code{\link{saveCombineArchive}()},
\code{\link{saveModelToString}()},
\code{\link{saveModel}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
