% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newCompartment}
\alias{newCompartment}
\title{Create a new compartment}
\usage{
newCompartment(name, type = c("fixed", "assignment", "ode"),
  initial_size = NULL, initial_expression = NULL, expression = NULL,
  model = getCurrentModel())
}
\arguments{
\item{name}{string}

\item{type}{string}

\item{initial_size}{number}

\item{initial_expression}{string}

\item{expression}{string}

\item{model}{a model object}
}
\description{
\code{newCompartment} creates a new compartment.
}
\details{
Default initial size is 1.
Arguments priority from lowest to highest is \code{initial_size}, \code{initial_expression}.
}
\seealso{
Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{entity_finders}},
  \code{\link{getCompartmentReferences}},
  \code{\link{getCompartments}},
  \code{\link{setCompartments}}
}
\concept{compartment functions}
