% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newCompartment}
\alias{newCompartment}
\title{Create a new compartment}
\usage{
newCompartment(name, type = c("fixed", "assignment", "ode"),
  initial.volume = 1, expression = NULL, model = getCurrentModel())
}
\arguments{
\item{name}{string}

\item{type}{string}

\item{initial.volume}{number}

\item{expression}{string}

\item{model}{a model object}
}
\description{
Create a new compartment
}
