% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newGlobalQuantity}
\alias{newGlobalQuantity}
\title{Create a new global quantity}
\usage{
newGlobalQuantity(name, type = c("fixed", "assignment", "ode"),
  unit = NULL, initial_value = NULL, initial_expression = NULL,
  expression = NULL, model = getCurrentModel())
}
\arguments{
\item{name}{string}

\item{type}{string}

\item{unit}{string}

\item{initial_value}{number}

\item{initial_expression}{Initial expression to set, as string, finite numeric, or logical.}

\item{expression}{Expression to set, as string, finite numeric, or logical.}

\item{model}{a model object}
}
\value{
quantity key
}
\description{
\code{newGlobalQuantity} creates a new global quantity.
}
\details{
Arguments priority from lowest to highest is \code{initial_value}, \code{initial_expression}.

The \href{https://jpahle.github.io/CoRC/articles/model_building.html}{online article on model building} provides some further context.
}
\seealso{
Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{setGlobalQuantities}}
}
\concept{global quantity functions}
