% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newKineticFunction}
\alias{newKineticFunction}
\title{Create a new kinetic function}
\usage{
newKineticFunction(name, formula, parameters, function.type = c("general",
  "reversible", "irreversible"))
}
\arguments{
\item{name}{string}

\item{formula}{string}

\item{parameters}{named character vector

Allowed values: substrate, product, modifier, parameter, volume, time}

\item{function.type}{string}
}
\value{
function key
}
\description{
Create a new kinetic function
}
\seealso{
Other reaction functions: \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
