% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newKineticFunction}
\alias{newKineticFunction}
\title{Create a new kinetic function}
\usage{
newKineticFunction(name, formula, parameters, function_type = c("general",
  "reversible", "irreversible"))
}
\arguments{
\item{name}{Name to set, as string}

\item{formula}{Function formula to set, as string}

\item{parameters}{Parameter types to set, as named character vector.
Defaults to 'parameter' for every parameter.

Allowed values: 'substrate', 'product', 'modifier', 'parameter', 'volume', 'time'.}

\item{function_type}{Type of function to set, as string.
Defines if the function is gerneral, reversible or irreversible.}
}
\value{
function key
}
\description{
Create a new kinetic function
}
\seealso{
Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
