% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newReaction}
\alias{newReaction}
\title{Create a new reaction}
\usage{
newReaction(
  reaction,
  name = reaction,
  fun = NULL,
  mappings = NULL,
  model = getCurrentModel()
)
}
\arguments{
\item{reaction}{Reaction equation to set, as string.}

\item{name}{Name to set, as string.
Defaults to reaction.}

\item{fun}{Identify which kinetic function to set by specifying it's key, as string.}

\item{mappings}{Parameter mappings, as named list.
Names are the parameters of the kinetic function.
Values are either entity keys valid for the specific type of parameter or a numeric value in case of a local parameter.}

\item{model}{a model object}
}
\value{
reaction key
}
\description{
\code{newReaction} creates a new reaction.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/model_building.html}{online article on model building} provides some further context.
}
\seealso{
Other reaction functions: 
\code{\link{clearCustomKineticFunctions}()},
\code{\link{deleteKineticFunction}()},
\code{\link{deleteReaction}()},
\code{\link{entity_finders}},
\code{\link{getParameterReferences}()},
\code{\link{getParameters}()},
\code{\link{getReactionMappings}()},
\code{\link{getReactionReferences}()},
\code{\link{getReactions}()},
\code{\link{getValidReactionFunctions}()},
\code{\link{newKineticFunction}()},
\code{\link{setParameters}()},
\code{\link{setReactionFunction}()},
\code{\link{setReactionMappings}()},
\code{\link{setReactions}()}
}
\concept{reaction functions}
