% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{newSpecies}
\alias{newSpecies}
\title{Create a new species}
\usage{
newSpecies(name, compartment = NULL, type = c("fixed", "assignment",
  "reactions", "ode"), initial.concentration = 1, expression = NULL,
  model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{newSpecies} creates a new species.
}
