% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{openCopasi}
\alias{openCopasi}
\title{Open the given model in the copasi UI}
\usage{
openCopasi(readin = FALSE, copasi_loc = "CopasiUI",
  datamodel = getCurrentModel())
}
\arguments{
\item{readin}{if \code{TRUE}, the function waits for Copasi to quit and then reads in the temporary model file, overwriting the give datamodel}

\item{copasi_loc}{location of CopasiUI}

\item{datamodel}{a model object}
}
\description{
Open the given model in the copasi UI
}
