% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{parameter}
\alias{parameter}
\title{Identify reaction parameter by name}
\usage{
parameter(key, reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a vector of strings to identify reaction parameters}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{model}{a model object}
}
\value{
a character vector of copasi keys
}
\description{
\code{parameter} identifies reaction parameters from the given name fragments.
}
