% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{quantity}
\alias{quantity}
\title{Identify global quantities by name}
\usage{
quantity(key = "", reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a string to identify global quantities.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of global quantity identifiers or references.
}
\description{
\code{quantity} identifies global quantities matching a given name fragment.
}
