% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_finders.R
\name{reaction}
\alias{reaction}
\title{Identify reactions by name}
\usage{
reaction(key = "", reference = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a string to identify reactions.}

\item{reference}{an optional string naming the value references to be returned.}

\item{model}{a model object.}
}
\value{
a character vector of reaction identifiers or references.
}
\description{
\code{reaction} identifies reactions matching a given name fragment.
}
