% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stringr.R
\docType{import}
\name{regex}
\alias{regex}
\alias{reexports}
\alias{fixed}
\alias{coll}
\title{String matching engines from stringr}
\details{
CoRC makes use of the string matching engines from the \pkg{stringr} package.

The functions are described at \link[stringr]{modifiers}.
By default, matching is done with \code{coll}.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stringr}{\code{\link[stringr:modifiers]{coll}}, \code{\link[stringr:modifiers]{fixed}}, \code{\link[stringr:modifiers]{regex}}}
}}

