% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lna.R
\name{runLinearNoiseApproximation}
\alias{runLinearNoiseApproximation}
\alias{runLNA}
\title{Run linear noise approximation}
\usage{
runLinearNoiseApproximation(performSteadyStateAnalysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

runLNA(performSteadyStateAnalysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{performSteadyStateAnalysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runLinearNoiseApproximation} runs linear noise approximation and returns the results in a list.
}
\seealso{
Other linear noise approximation: \code{\link{getLinearNoiseApproximationSettings}},
  \code{\link{setLinearNoiseApproximationSettings}}
}
\concept{linear noise approximation}
