% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_optimization.R
\name{runOptimization}
\alias{runOptimization}
\alias{runOpt}
\title{Run optimization}
\usage{
runOptimization(expression = NULL, maximize = NULL, subtask = NULL,
  randomize_start_values = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())

runOpt(expression = NULL, maximize = NULL, subtask = NULL,
  randomize_start_values = NULL, calculate_statistics = NULL,
  update_model = NULL, executable = NULL, parameters = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{expression}{Expression to optimize, as string.}

\item{maximize}{flag}

\item{subtask}{string

Available tasks: 'steadyState', 'timeCourse', 'scan', 'fluxMode', 'optimization', 'parameterFitting', 'mca', 'lyap', 'tssAnalysis', 'sens', 'moieties', 'crosssection', 'lna', 'analytics'.}

\item{randomize_start_values}{flag}

\item{calculate_statistics}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{parameters}{corc_opt_parm or list of corc_opt_parm objects}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'Statistics', 'DifferentialEvolution', 'SRES', 'EvolutionaryProgram', 'GeneticAlgorithm', 'GeneticAlgorithmSR', 'HookeJeeves', 'LevenbergMarquardt', 'NelderMead', 'ParticleSwarm', 'Praxis', 'RandomSearch', 'ScatterSearch', 'SimulatedAnnealing', 'SteepestDescent', 'TruncatedNewton'.}

\item{model}{A model object.}
}
\value{
A list of results.
}
\description{
\code{runOptimization} runs optimization and returns the results in a list.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other optimization: \code{\link{addOptimizationParameter}},
  \code{\link{clearOptimizationParameters}},
  \code{\link{defineOptimizationParameter}},
  \code{\link{getOptimizationSettings}},
  \code{\link{setOptimizationSettings}}
}
\concept{optimization}
