% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{runParameterEstimation}
\alias{runParameterEstimation}
\alias{runPE}
\title{Run parameter estimation}
\usage{
runParameterEstimation(randomizeStartValues = NULL,
  createParameterSets = NULL, calculateStatistics = NULL,
  updateModel = NULL, executable = NULL, parameters = NULL,
  experiments = NULL, method = NULL, model = getCurrentModel())

runPE(randomizeStartValues = NULL, createParameterSets = NULL,
  calculateStatistics = NULL, updateModel = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{randomizeStartValues}{flag}

\item{createParameterSets}{flag}

\item{calculateStatistics}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{parameters}{copasi_param or list of copasi_param objects}

\item{experiments}{copasi_exp or list of copasi_exp objects}

\item{method}{string or list}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runParameterEstimation} runs parameter estimation and returns the results in a list.
}
