% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{runParameterEstimation}
\alias{runParameterEstimation}
\alias{runPE}
\title{Run parameter estimation}
\usage{
runParameterEstimation(
  randomize_start_values = NULL,
  create_parameter_sets = NULL,
  calculate_statistics = NULL,
  use_time_sens = NULL,
  update_model = NULL,
  executable = NULL,
  parameters = NULL,
  experiments = NULL,
  method = NULL,
  model = getCurrentModel()
)

runPE(
  randomize_start_values = NULL,
  create_parameter_sets = NULL,
  calculate_statistics = NULL,
  use_time_sens = NULL,
  update_model = NULL,
  executable = NULL,
  parameters = NULL,
  experiments = NULL,
  method = NULL,
  model = getCurrentModel()
)
}
\arguments{
\item{randomize_start_values}{flag}

\item{create_parameter_sets}{flag}

\item{calculate_statistics}{flag}

\item{use_time_sens}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{parameters}{corc_opt_parm or list of corc_opt_parm objects

See also \code{\link{defineParameterEstimationParameter}}.}

\item{experiments}{copasi_exp or list of copasi_exp objects

See also \code{\link{defineExperiments}}.}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'Statistics', 'DifferentialEvolution', 'SRES', 'EvolutionaryProgram', 'GeneticAlgorithm', 'GeneticAlgorithmSR', 'HookeJeeves', 'LevenbergMarquardt', 'NL2SOL', 'NelderMead', 'ParticleSwarm', 'Praxis', 'RandomSearch', 'ScatterSearch', 'SimulatedAnnealing', 'SteepestDescent', 'TruncatedNewton'.}

\item{model}{A model object.}
}
\value{
A list of results.
}
\description{
\code{runParameterEstimation} runs parameter estimation and returns the results in a list.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other parameter estimation: 
\code{\link{addExperiments}()},
\code{\link{addParameterEstimationParameter}()},
\code{\link{clearExperiments}()},
\code{\link{clearParameterEstimationParameters}()},
\code{\link{clearValidations}()},
\code{\link{defineExperiments}()},
\code{\link{getParameterEstimationSettings}()},
\code{\link{setParameterEstimationSettings}()}
}
\concept{parameter estimation}
