% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{runSteadyState}
\alias{runSteadyState}
\alias{runSS}
\title{Run to stready state}
\usage{
runSteadyState(calculateJacobian = NULL, performStabilityAnalysis = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runSS(calculateJacobian = NULL, performStabilityAnalysis = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{calculateJacobian}{flag}

\item{performStabilityAnalysis}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runSteadyState} calculates the steady state and returns the results in a list.
}
