% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{runSteadyState}
\alias{runSteadyState}
\title{Run to stready state}
\usage{
runSteadyState(calculateJacobian = NULL, performStabilityAnalysis = NULL,
  updateModel = NULL, method = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{calculateJacobian}{boolean}

\item{performStabilityAnalysis}{boolean}

\item{updateModel}{boolean}

\item{method}{list}

\item{datamodel}{a model object}
}
\value{
a list of results
}
\description{
\code{runSteadyState} calculates the steady state and returns results in a list.
}
