% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{runSteadyState}
\alias{runSteadyState}
\alias{runSS}
\title{Run to steady state}
\usage{
runSteadyState(calculate_jacobian = NULL, perform_stability_analysis = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runSS(calculate_jacobian = NULL, perform_stability_analysis = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{calculate_jacobian}{flag}

\item{perform_stability_analysis}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{A model object.}
}
\value{
A list of results.
\itemize{
  \item \code{$result} can be one of 'notFound', 'found', 'foundEquilibrium', 'foundNegative'.
}
}
\description{
\code{runSteadyState} calculates the steady state and returns the results in a list.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other steady state: \code{\link{getSteadyStateSettings}},
  \code{\link{setSteadyStateSettings}}
}
\concept{steady state}
