% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimeCourse}
\alias{runTimeCourse}
\alias{runTC}
\title{Run a time course}
\usage{
runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runTC(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{duration}{number}

\item{dt}{number}

\item{intervals}{count}

\item{suppress_output_before}{flag}

\item{output_events}{flag}

\item{save_result_in_memory}{flag}

\item{start_in_steady_state}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method args and their values.

Allowed methods: 'deterministic', 'stochastic', 'directMethod', 'tauLeap', 'adaptiveSA', 'hybrid', 'hybridLSODA', 'hybridODE45'.}

\item{model}{A model object.}
}
\value{
A list of results.
}
\description{
\code{runTimeCourse} runs a time course and returns the results in a list.
}
\seealso{
Other time course: \code{\link{getTimeCourseSettings}},
  \code{\link{setTimeCourseSettings}}
}
\concept{time course}
