% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimeCourse}
\alias{runTimeCourse}
\alias{runTC}
\title{Run a time course}
\usage{
runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runTC(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{duration}{number}

\item{dt}{number}

\item{intervals}{count}

\item{suppressOutputBefore}{flag}

\item{outputEvents}{flag}

\item{saveResultInMemory}{flag}

\item{startInSteadyState}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{method}{string or list}

\item{model}{a model object}
}
\value{
A data frame with a time column and value columns.
  Has attributes for units and settings.
}
\description{
\code{runTimeCourse} runs a time course and returns the time course data as a data frame.
}
