% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimeCourse}
\alias{runTimeCourse}
\alias{runTC}
\title{Run a time course}
\usage{
runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

runTC(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{duration}{number}

\item{dt}{number}

\item{intervals}{count}

\item{suppressOutputBefore}{flag}

\item{outputEvents}{flag}

\item{saveResultInMemory}{flag}

\item{startInSteadyState}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{method}{string or list

Set method with a string or a list with the entry `method`. The list may contain further method args and their values.
Allowed methods: deterministic, stochastic, directMethod, tauLeap, adaptiveSA, hybrid, hybridLSODA, hybridODE45}

\item{model}{a model object}
}
\value{
a list of results
}
\description{
\code{runTimeCourse} runs a time course and returns the results in a list.
}
\seealso{
Other time course: \code{\link{getTimeCourseSettings}},
  \code{\link{setTimeCourseSettings}}
}
\concept{time course}
