% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{runTimeCourse}
\alias{runTimeCourse}
\alias{runTC}
\title{Run a time course}
\usage{
runTimeCourse(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  soft_error = FALSE, model = getCurrentModel())

runTC(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  soft_error = FALSE, model = getCurrentModel())
}
\arguments{
\item{duration}{The time course duration, as number.}

\item{dt}{The time course output step size, as number.}

\item{intervals}{The time course step count, as count.
Overwrites \code{dt} in case of conflict.}

\item{suppress_output_before}{Whether to suppress before a certain time point, as number.}

\item{output_events}{Whether to output events as additional steps, as flag.}

\item{save_result_in_memory}{Whether to generate an output data frame for the time course, as flag.}

\item{start_in_steady_state}{Whether to first go to steady state before running the time course, as flag.}

\item{update_model}{Whether to update the model with the state resulting from the time course, as flag.}

\item{executable}{flag}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'deterministic', 'stochastic', 'directMethod', 'tauLeap', 'adaptiveSA', 'hybrid', 'hybridLSODA', 'hybridODE45', 'stochasticRunkeKuttaRI5'.}

\item{soft_error}{Whether to divert task processing errors to the \code{task_error} result entry, as flag.
Allows for readout of incomplete result tables for failed time courses.}

\item{model}{A model object.}
}
\value{
A list of results.
\itemize{
  \item \code{$result} is a data frame containing time course output as concentrations.
  \item \code{$result_number} is a data frame containing time course output as particle numbers.
  \item \code{$task_error} is \code{NULL} or, in case \code{soft_error} is set to \code{TRUE} and the time course failed, a string giving an error message.
}
\strong{Attention}: Both result data frames will be empty if \code{save_result_in_memory} is set to \code{FALSE}.
Test the current value of this parameter with: \code{getTimeCourseSettings()$save_result_in_memory}.
}
\description{
\code{runTimeCourse} runs a time course and returns the results in a list.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other time course: \code{\link{getTimeCourseSettings}},
  \code{\link{setTimeCourseSettings}}
}
\concept{time course}
