% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveModel}
\alias{saveModel}
\title{Save the model as a .cps file}
\usage{
saveModel(
  filename = model$getFileName(),
  overwrite = FALSE,
  model = getCurrentModel()
)
}
\arguments{
\item{filename}{a path to save to}

\item{overwrite}{is overwriting existing files allowed?}

\item{model}{a model object}
}
\description{
\code{saveModel} saves the given model as a .cps file
}
\seealso{
Other model loading: 
\code{\link{loadExamples}()},
\code{\link{loadModelFromString}()},
\code{\link{loadModel}()},
\code{\link{loadSBMLFromString}()},
\code{\link{loadSBML}()},
\code{\link{newModel}()},
\code{\link{saveModelToString}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
