% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{saveModelToString}
\alias{saveModelToString}
\title{Save the model to string}
\usage{
saveModelToString(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\value{
The COPASI model file, as string.
}
\description{
\code{saveModelToString} returns the given model as a string.
}
\seealso{
Other model loading: 
\code{\link{loadExamples}()},
\code{\link{loadModelFromString}()},
\code{\link{loadModel}()},
\code{\link{loadSBMLFromString}()},
\code{\link{loadSBML}()},
\code{\link{newModel}()},
\code{\link{saveModel}()},
\code{\link{saveSBMLToString}()},
\code{\link{saveSBML}()},
\code{\link{unloadAllModels}()},
\code{\link{unloadModel}()}
}
\concept{model loading}
