% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setCompartments}
\alias{setCompartments}
\title{Set Compartments}
\usage{
setCompartments(compartments, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{compartments}{a data frame as given by getCompartments()}

\item{datamodel}{a model object}
}
\description{
\code{setCompartments} accepts a data frame of compartments and attempts to apply given values to the model depending on the 'key' column.
}
