% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setCompartments}
\alias{setCompartments}
\title{Set compartments}
\usage{
setCompartments(key = NULL, name = NULL, initial.volume = NULL,
  data = NULL, datamodel = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying compartments}

\item{name}{a character vector of names to set}

\item{initial.volume}{a numeric vector of values to set}

\item{data}{a data frame as given by \code{getCompartments} which will be applied before the other arguments.}

\item{datamodel}{a model object}
}
\description{
\code{setCompartments} applies given values to compartments of the model depending on the 'key' parameter.
}
