% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setCompartments}
\alias{setCompartments}
\title{Set compartments}
\usage{
setCompartments(key = NULL, name = NULL, type = NULL,
  dimensionality = NULL, initial_size = NULL, initial_expression = NULL,
  expression = NULL, data = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{Identify which compartment to edit by specifying it's key, as string.
Also supports fragments of keys, if uniquely matching one compartment.}

\item{name}{Name to set, as string.}

\item{type}{Type ("fixed", "assignment", "ode") to set, as string.}

\item{dimensionality}{Dimensionality to set (0D, 1D, 2D, 3D), as number.}

\item{initial_size}{Initial size to set, as string.}

\item{initial_expression}{Initial expression to set, as string, finite numeric, or logical.}

\item{expression}{Expression to set, as string, finite numeric, or logical.}

\item{data}{A data frame as given by \code{\link{getCompartments}} which will be applied before the other arguments.}

\item{model}{A model object.}
}
\description{
\code{setCompartments} applies given values to compartments of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which compartment to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.

The \href{https://jpahle.github.io/CoRC/articles/entity_management.html}{online article on managing model entities} provides some further context.
}
\seealso{
\code{\link{getCompartments}} \code{\link{getCompartmentReferences}}

Other compartment functions: \code{\link{deleteCompartment}},
  \code{\link{entity_finders}},
  \code{\link{getCompartmentReferences}},
  \code{\link{getCompartments}},
  \code{\link{newCompartment}}
}
\concept{compartment functions}
