% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{setCurrentModel}
\alias{setCurrentModel}
\title{Set the currently active model}
\usage{
setCurrentModel(datamodel)
}
\arguments{
\item{datamodel}{a model object}
}
\description{
\code{setCurrentModel} sets the given model as the currently active model.
}
