% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setGlobalQuantities}
\alias{setGlobalQuantities}
\title{Set global quantities}
\usage{
setGlobalQuantities(key = NULL, name = NULL, type = NULL,
  initial_value = NULL, expression = NULL, data = NULL,
  model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{name}{a character vector of names to set}

\item{type}{a character vector of types ("fixed", "assignment", "ode").}

\item{expression}{a character vector of expressions to set}

\item{data}{a data frame as given by \code{\link{getGlobalQuantities}} which will be applied before the other arguments.}

\item{model}{a model object}

\item{initial.volume}{a numeric vector of values to set}
}
\description{
\code{setGlobalQuantities} applies given values to global quantities of the model depending on the 'key' parameter.
}
\seealso{
\code{\link{getGlobalQuantities}} \code{\link{getGlobalQuantityReferences}}

Other global quantity functions: \code{\link{deleteGlobalQuantity}},
  \code{\link{entity_finders}},
  \code{\link{getGlobalQuantities}},
  \code{\link{getGlobalQuantityReferences}},
  \code{\link{newGlobalQuantity}}
}
\concept{global quantity functions}
