% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setGlobalQuantities}
\alias{setGlobalQuantities}
\title{Set global quantities}
\usage{
setGlobalQuantities(quantities, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{quantities}{a data frame as given by getGlobalQuantities()}

\item{datamodel}{a model object}
}
\description{
\code{setGlobalQuantities} accepts a data frame of global quantities and attempts to apply given values to the model depending on the 'key' column.
}
