% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setGlobalQuantities}
\alias{setGlobalQuantities}
\title{Set Global Quantities}
\usage{
setGlobalQuantities(quantities, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{datamodel}{a model object}

\item{species}{a data frame as given from getGlobalQuantities()}
}
\description{
\code{setGlobalQuantities} accepts a data frame of global quantities and attempts to apply given values to the model depending on the 'key' column.
}
