% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setGlobalQuantities}
\alias{setGlobalQuantities}
\title{Set global quantities}
\usage{
setGlobalQuantities(key = NULL, name = NULL, initial.value = NULL,
  data = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying global quantities}

\item{name}{a character vector of names to set}

\item{data}{a data frame as given by \code{getGlobalQuantities} which will be applied before the other arguments.}

\item{datamodel}{a model object}

\item{initial.volume}{a numeric vector of values to set}
}
\description{
\code{setGlobalQuantities} applies given values to global quantities of the model depending on the 'key' parameter.
}
