% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setInitialTime}
\alias{setInitialTime}
\title{Set the model's initial time}
\usage{
setInitialTime(time, model = getCurrentModel())
}
\arguments{
\item{time}{numeric}

\item{model}{a model object}
}
\description{
\code{setInitialTime} sets the initial time of the model.
}
\seealso{
\code{\link{getInitialTime}}

Other model property functions: \code{\link{getAreaUnit}},
  \code{\link{getInitialTime}},
  \code{\link{getLengthUnit}}, \code{\link{getLinkMatrix}},
  \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setAreaUnit}}, \code{\link{setLengthUnit}},
  \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}},
  \code{\link{setVolumeUnit}}
}
\concept{model property functions}
