% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_lna.R
\name{setLinearNoiseApproximationSettings}
\alias{setLinearNoiseApproximationSettings}
\alias{setLNA}
\title{Set linear noise approximation settings}
\usage{

  setLinearNoiseApproximationSettings(perform_steady_state_analysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

setLNA(perform_steady_state_analysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{perform_steady_state_analysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\description{
\code{setLinearNoiseApproximation} sets linear noise approximation task settings including method options.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other linear noise approximation: \code{\link{getLinearNoiseApproximationSettings}},
  \code{\link{runLinearNoiseApproximation}}
}
\concept{linear noise approximation}
