% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_mca.R
\name{setMetabolicControlAnalysisSettings}
\alias{setMetabolicControlAnalysisSettings}
\alias{setMCA}
\title{Set metabolic control analysis settings}
\usage{
setMetabolicControlAnalysisSettings(performSteadyStateAnalysis = NULL,
  executable = NULL, method = NULL, model = getCurrentModel())

setMCA(performSteadyStateAnalysis = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())
}
\arguments{
\item{performSteadyStateAnalysis}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\description{
\code{setMetabolicControlAnalysis} sets metabolic control analysis task settings including method options.
}
\seealso{
Other metabolic control analysis: \code{\link{getMetabolicControlAnalysisSettings}},
  \code{\link{runMetabolicControlAnalysis}}
}
\concept{metabolic control analysis}
