% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_properties.R
\name{setModelName}
\alias{setModelName}
\title{Set a model's name}
\usage{
setModelName(name, datamodel = getCurrentModel())
}
\arguments{
\item{name}{the new model name}

\item{datamodel}{a model object}
}
\description{
\code{setModelName} sets the name of the model.
}
