% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parameterestimation.R
\name{setParamEstSettings}
\alias{setParamEstSettings}
\title{Set parameter estimation settings}
\usage{
setParamEstSettings(randomizeStartValues = NULL, createParameterSets = NULL,
  calculateStatistics = NULL, updateModel = NULL, executable = NULL,
  parameters = NULL, experiments = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{randomizeStartValues}{boolean}

\item{createParameterSets}{boolean}

\item{calculateStatistics}{boolean}

\item{updateModel}{boolean}

\item{executable}{boolean}

\item{method}{character or list}

\item{model}{a model object}
}
\description{
\code{setParamEstSettings} sets parameter estimation settings including method options.
}
