% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setParameters}
\alias{setParameters}
\title{Set reaction parameters}
\usage{
setParameters(key = NULL, name = NULL, value = NULL, data = NULL,
  datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying reaction parameters}

\item{name}{a character vector of names to set}

\item{data}{a data frame as given by getParameters which will be applied before the other arguments.}

\item{datamodel}{a model object}
}
\description{
\code{setParameters} applies given values to reaction parameters of the model depending on the 'key' parameter.
}
