% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setParameters}
\alias{setParameters}
\title{Set reaction parameters}
\usage{
setParameters(key = NULL, name = NULL, value = NULL,
  mapping = NULL, data = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{Identify which reaction parameter to edit by specifying it's key, as string.
Also supports fragments of keys, if uniquely matching one reaction parameter.}

\item{name}{Name to set, as string.}

\item{value}{Value to set, as numeric.}

\item{mapping}{Key of global quantity to map to, as string.
Also supports fragments of keys, if uniquely matching one global quantity.}

\item{data}{A data frame as given by \code{\link{getParameters}} which will be applied before the other arguments.}

\item{model}{A model object.}
}
\description{
\code{setParameters} applies given values to reaction parameters of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which reaction parameter to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.

The \href{https://jpahle.github.io/CoRC/articles/entity_management.html}{online article on managing model entities} provides some further context.
}
\seealso{
\code{\link{getParameters}} \code{\link{getParameterReferences}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
