% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setQuantityUnit}
\alias{setQuantityUnit}
\title{Set the unit of quantity}
\usage{
setQuantityUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setQuantityUnit} sets the unit used for quantity.
}
\seealso{
\code{\link{getQuantityUnit}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setTimeUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
