% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setReactionFunction}
\alias{setReactionFunction}
\title{Set a reaction function}
\usage{
setReactionFunction(key, fun, mappings = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{fun}{string}

\item{mappings}{named list}

\item{model}{a model object}
}
\description{
Set a reaction function
}
\seealso{
\code{\link{getValidReactionFunctions}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionMappings}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
