% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setReactionMappings}
\alias{setReactionMappings}
\title{Set reaction parameter mappings}
\usage{
setReactionMappings(key, mappings, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{mappings}{list}

\item{model}{a model object}
}
\description{
Set reaction parameter mappings
}
