% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setReactionMappings}
\alias{setReactionMappings}
\title{Set reaction parameter mappings}
\usage{
setReactionMappings(key, mappings, model = getCurrentModel())
}
\arguments{
\item{key}{reaction key}

\item{mappings}{list}

\item{model}{a model object}
}
\description{
Set reaction parameter mappings
}
\seealso{
\code{\link{getReactionMappings}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactions}}
}
\concept{reaction functions}
