% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setReactions}
\alias{setReactions}
\title{Set reactions}
\usage{
setReactions(key = NULL, name = NULL, data = NULL,
  datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying reactions}

\item{name}{a character vector of names to set}

\item{data}{a data frame as given by \code{getReactions} which will be applied before the other arguments.}

\item{datamodel}{a model object}
}
\description{
\code{setReactions} applies given values to reactions of the model depending on the 'key' parameter.
}
