% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setReactions}
\alias{setReactions}
\title{Set Reactions}
\usage{
setReactions(reactions, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{reactions}{a data frame as given by getReactions()}

\item{datamodel}{a model object}
}
\description{
\code{setReactions} accepts a data frame of reactions and attempts to apply given values to the model depending on the 'key' column.
}
