% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setReactions}
\alias{setReactions}
\title{Set reactions}
\usage{
setReactions(key = NULL, name = NULL, data = NULL,
  model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying reactions}

\item{name}{a character vector of names to set}

\item{data}{a data frame as given by \code{\link{getReactions}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setReactions} applies given values to reactions of the model depending on the 'key' parameter.
}
\seealso{
\code{\link{getReactions}} \code{\link{getReactionReferences}}

Other reaction functions: \code{\link{clearCustomKineticFunctions}},
  \code{\link{deleteKineticFunction}},
  \code{\link{deleteReaction}},
  \code{\link{entity_finders}},
  \code{\link{getParameterReferences}},
  \code{\link{getParameters}},
  \code{\link{getReactionMappings}},
  \code{\link{getReactionReferences}},
  \code{\link{getReactions}},
  \code{\link{getValidReactionFunctions}},
  \code{\link{newKineticFunction}},
  \code{\link{newReaction}}, \code{\link{setParameters}},
  \code{\link{setReactionFunction}},
  \code{\link{setReactionMappings}}
}
\concept{reaction functions}
