% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(species, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{species}{a data frame as given from getSpecies()}

\item{datamodel}{a model object}
}
\description{
\code{setSpecies} accepts a data frame of species and attempts to apply given values to the model depending on the 'key' column.
}
