% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(key = NULL, name = NULL, concentration = NULL,
  particlenum = NULL, data = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{name}{a character vector of names to set}

\item{concentration}{a character vector of concentrations to set}

\item{particlenum}{a character vector of particle numbers to set}

\item{data}{a data frame as given by getSpecies which will be applied before the other arguments}

\item{datamodel}{a model object}
}
\description{
\code{setSpecies} accepts a data frame of species and attempts to apply given values to the model depending on the 'key' column.
}
