% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(key = NULL, name = NULL, compartment = NULL, type = NULL,
  initial_concentration = NULL, initial_number = NULL,
  initial_expression = NULL, expression = NULL, data = NULL,
  model = getCurrentModel())
}
\arguments{
\item{key}{Identify which species to edit by specifying it's key, as string.
Also supports fragments of keys, if uniquely matching one species.}

\item{name}{Name to set, as string.}

\item{compartment}{Key of new compartment to set, as string.}

\item{type}{Type ("fixed", "assignment", "reactions", "ode") to set, as string.}

\item{initial_concentration}{Initial concentration to set, as numeric.}

\item{initial_number}{Initial particle number to set, as numeric.}

\item{initial_expression}{Initial expression to set, as string, finite numeric, or logical.}

\item{expression}{Expression to set, as string, finite numeric, or logical.}

\item{data}{A data frame as given by \code{\link{getSpecies}} which will be applied before the other arguments.}

\item{model}{A model object.}
}
\description{
\code{setSpecies} applies given values to species of the model depending on the \code{key} argument.
}
\details{
Use the \code{key} argument to specify which species to modify and any of the other arguments to specify the value to set.
The function is fully vectorized.
If a \code{NA} value is supplied, the model value is kept unchanged.
}
\seealso{
\code{\link{getSpecies}} \code{\link{getSpeciesReferences}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{newSpecies}}
}
\concept{species functions}
