% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(key = NULL, name = NULL, type = NULL,
  initial.concentration = NULL, initial.number = NULL, expression = NULL,
  data = NULL, model = getCurrentModel())
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{name}{a character vector of names to set}

\item{type}{a character vector of types ("fixed", "assignment", "reactions", "ode").}

\item{initial.concentration}{a numeric vector of concentrations to set}

\item{initial.number}{a numeric vector of particle numbers to set}

\item{expression}{a character vector of expressions to set}

\item{data}{a data frame as given by \code{\link{getSpecies}} which will be applied before the other arguments.}

\item{model}{a model object}
}
\description{
\code{setSpecies} applies given values to species of the model depending on the 'key' parameter.
}
\seealso{
\code{\link{getSpecies}} \code{\link{getSpeciesReferences}}

Other species functions: \code{\link{deleteSpecies}},
  \code{\link{entity_finders}},
  \code{\link{getSpeciesReferences}},
  \code{\link{getSpecies}}, \code{\link{newSpecies}}
}
\concept{species functions}
