% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_accessors.R
\name{setSpecies}
\alias{setSpecies}
\title{Set species}
\usage{
setSpecies(key = NULL, name = NULL, initial.concentration = NULL,
  initial.number = NULL, data = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a character vector uniquely identifying species}

\item{name}{a character vector of names to set}

\item{initial.concentration}{a numeric vector of concentrations to set}

\item{initial.number}{a numeric vector of particle numbers to set}

\item{data}{a data frame as given by \code{getSpecies} which will be applied before the other arguments.}

\item{datamodel}{a model object}
}
\description{
\code{setSpecies} applies given values to species of the model depending on the 'key' parameter.
}
