% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{setSteadyStateSettings}
\alias{setSteadyStateSettings}
\title{Set steady state settings}
\usage{
setSteadyStateSettings(calculateJacobian = NULL,
  performStabilityAnalysis = NULL, updateModel = NULL, method = NULL,
  executable = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{calculateJacobian}{boolean}

\item{performStabilityAnalysis}{boolean}

\item{updateModel}{boolean}

\item{method}{list}

\item{executable}{boolean}

\item{datamodel}{a model object}
}
\description{
\code{setSteadyStateSettings} sets steady state settings including method options.
}
