% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_steadystate.R
\name{setSteadyStateSettings}
\alias{setSteadyStateSettings}
\alias{setSS}
\title{Set steady state settings}
\usage{
setSteadyStateSettings(calculateJacobian = NULL,
  performStabilityAnalysis = NULL, updateModel = NULL, executable = NULL,
  method = NULL, model = getCurrentModel())

setSS(calculateJacobian = NULL, performStabilityAnalysis = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{calculateJacobian}{flag}

\item{performStabilityAnalysis}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{method}{list}

\item{model}{a model object}
}
\description{
\code{setSteadyStateSettings} sets steady state task settings including method options.
}
