% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{setTimeCourseSettings}
\alias{setTimeCourseSettings}
\alias{setTC}
\title{Set time course settings}
\usage{
setTimeCourseSettings(
  duration = NULL,
  dt = NULL,
  intervals = NULL,
  automatic_intervals = NULL,
  suppress_output_before = NULL,
  output_events = NULL,
  save_result_in_memory = NULL,
  start_in_steady_state = NULL,
  update_model = NULL,
  executable = NULL,
  method = NULL,
  model = getCurrentModel()
)

setTC(
  duration = NULL,
  dt = NULL,
  intervals = NULL,
  automatic_intervals = NULL,
  suppress_output_before = NULL,
  output_events = NULL,
  save_result_in_memory = NULL,
  start_in_steady_state = NULL,
  update_model = NULL,
  executable = NULL,
  method = NULL,
  model = getCurrentModel()
)
}
\arguments{
\item{duration}{The time course duration, as number.}

\item{dt}{The time course output step size, as number.}

\item{intervals}{The time course step count, as count.
Overwrites \code{dt} in case of conflict.}

\item{automatic_intervals}{Whether to use automatic intervals, as flag.}

\item{suppress_output_before}{Whether to suppress before a certain time point, as number.}

\item{output_events}{Whether to output events as additional steps, as flag.}

\item{save_result_in_memory}{Whether to generate an output data frame for the time course, as flag.}

\item{start_in_steady_state}{Whether to first go to steady state before running the time course, as flag.}

\item{update_model}{Whether to update the model with the state resulting from the time course, as flag.}

\item{executable}{flag}

\item{method}{string or list

Set the method directly with a string or with a list containing the entry named `method`.

The list may contain further method arguments and their values.

Available methods: 'deterministic', 'RADAU5', 'stochastic', 'directMethod', 'tauLeap', 'adaptiveSA', 'hybrid', 'hybridLSODA', 'hybridODE45', 'stochasticRunkeKuttaRI5'.}

\item{model}{A model object.}
}
\description{
\code{setTimeCourseSettings} sets time course task settings including method options.
}
\details{
The \href{https://jpahle.github.io/CoRC/articles/task_management.html}{online article on managing tasks} provides some further context.
}
\seealso{
Other time course: 
\code{\link{getTimeCourseSettings}()},
\code{\link{runTimeCourse}()}
}
\concept{time course}
