% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{setTimeCourseSettings}
\alias{setTimeCourseSettings}
\alias{setTC}
\title{Set time course settings}
\usage{
setTimeCourseSettings(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

setTC(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{duration}{number}

\item{dt}{number}

\item{intervals}{count}

\item{suppressOutputBefore}{flag}

\item{outputEvents}{flag}

\item{saveResultInMemory}{flag}

\item{startInSteadyState}{flag}

\item{updateModel}{flag}

\item{executable}{flag}

\item{method}{string or list}

\item{model}{a model object}
}
\description{
\code{setTimeCourseSettings} sets time course task settings including method options.
}
