% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{setTimeCourseSettings}
\alias{setTimeCourseSettings}
\title{Set time course settings}
\usage{
setTimeCourseSettings(duration = NULL, dt = NULL, intervals = NULL,
  suppressOutputBefore = NULL, outputEvents = NULL,
  saveResultInMemory = NULL, startInSteadyState = NULL,
  updateModel = NULL, method = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{duration}{numeric time course duration}

\item{dt}{numeric}

\item{intervals}{integer}

\item{suppressOutputBefore}{boolean}

\item{outputEvents}{boolean}

\item{saveResultInMemory}{boolean}

\item{startInSteadyState}{boolean}

\item{updateModel}{not yet implemented}

\item{method}{character or list}

\item{datamodel}{a model object}
}
\description{
\code{setTimeCourseSettings} sets time course settings including method options.
}
