% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_timecourse.R
\name{setTimeCourseSettings}
\alias{setTimeCourseSettings}
\alias{setTC}
\title{Set time course settings}
\usage{
setTimeCourseSettings(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())

setTC(duration = NULL, dt = NULL, intervals = NULL,
  suppress_output_before = NULL, output_events = NULL,
  save_result_in_memory = NULL, start_in_steady_state = NULL,
  update_model = NULL, executable = NULL, method = NULL,
  model = getCurrentModel())
}
\arguments{
\item{duration}{number}

\item{dt}{number}

\item{intervals}{count}

\item{suppress_output_before}{flag}

\item{output_events}{flag}

\item{save_result_in_memory}{flag}

\item{start_in_steady_state}{flag}

\item{update_model}{flag}

\item{executable}{flag}

\item{method}{string or list

Set method with a string or a list with the entry `method`. The list may contain further method args and their values.
Allowed methods: deterministic, stochastic, directMethod, tauLeap, adaptiveSA, hybrid, hybridLSODA, hybridODE45}

\item{model}{a model object}
}
\description{
\code{setTimeCourseSettings} sets time course task settings including method options.
}
\seealso{
Other time course: \code{\link{getTimeCourseSettings}},
  \code{\link{runTimeCourse}}
}
\concept{time course}
