% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setTimeUnit}
\alias{setTimeUnit}
\title{Set the unit of time}
\usage{
setTimeUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setTimeUnit} sets the unit used for time.
}
