% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setTimeUnit}
\alias{setTimeUnit}
\title{Set the unit of time}
\usage{
setTimeUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setTimeUnit} sets the unit used for time.
}
\seealso{
\code{\link{getTimeUnit}}

Other model property functions: 
\code{\link{getAreaUnit}()},
\code{\link{getInitialTime}()},
\code{\link{getLengthUnit}()},
\code{\link{getLinkMatrix}()},
\code{\link{getModelName}()},
\code{\link{getQuantityUnit}()},
\code{\link{getReducedStoichiometryMatrix}()},
\code{\link{getStoichiometryMatrix}()},
\code{\link{getTimeUnit}()},
\code{\link{getVolumeUnit}()},
\code{\link{setAreaUnit}()},
\code{\link{setInitialTime}()},
\code{\link{setLengthUnit}()},
\code{\link{setModelName}()},
\code{\link{setQuantityUnit}()},
\code{\link{setVolumeUnit}()}
}
\concept{model property functions}
