% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_properties.R
\name{setVolumeUnit}
\alias{setVolumeUnit}
\title{Set the unit of volume}
\usage{
setVolumeUnit(unit, model = getCurrentModel())
}
\arguments{
\item{unit}{string}

\item{model}{a model object}
}
\description{
\code{setVolumeUnit} sets the unit used for volume.
}
\seealso{
\code{\link{getTimeUnit}}

Other model property functions: \code{\link{getInitialTime}},
  \code{\link{getLinkMatrix}}, \code{\link{getModelName}},
  \code{\link{getQuantityUnit}},
  \code{\link{getReducedStoichiometryMatrix}},
  \code{\link{getStoichiometryMatrix}},
  \code{\link{getTimeUnit}}, \code{\link{getVolumeUnit}},
  \code{\link{setInitialTime}}, \code{\link{setModelName}},
  \code{\link{setQuantityUnit}}, \code{\link{setTimeUnit}}
}
\concept{model property functions}
