% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_finders.R
\name{species}
\alias{species}
\title{Identify species by name}
\usage{
species(key, reference = NULL, datamodel = pkg_env$curr_dm)
}
\arguments{
\item{key}{a vector of strings to identify species}

\item{reference}{a scalar character or vector of characters naming the value references to be returned.}

\item{datamodel}{a model object}
}
\value{
a character vector of copasi keys
}
\description{
\code{species} identifies species from the given name fragments.
}
