% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_models.R
\name{unloadModel}
\alias{unloadModel}
\title{Unload a model}
\usage{
unloadModel(datamodel = getCurrentModel())
}
\arguments{
\item{datamodel}{a model object}
}
\description{
\code{unloadModel} frees memory by unloading a model from copasi
}
