% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{unloadModel}
\alias{unloadModel}
\title{Unload a model}
\usage{
unloadModel(model = getCurrentModel())
}
\arguments{
\item{model}{a model object}
}
\description{
\code{unloadModel} frees memory by unloading the given model from copasi
}
\seealso{
Other model loading: \code{\link{loadExamples}},
  \code{\link{loadModelFromString}},
  \code{\link{loadModel}},
  \code{\link{loadSBMLFromString}}, \code{\link{loadSBML}},
  \code{\link{newModel}}, \code{\link{saveModelToString}},
  \code{\link{saveModel}}, \code{\link{saveSBMLToString}},
  \code{\link{saveSBML}}, \code{\link{unloadAllModels}}
}
\concept{model loading}
